Migrate fixtures, move components that aren't really components into school_overview
