+++ app/assets/javascripts/district_enrollment/districtEnrollmentJson.js
