+++ app/assets/javascripts/educator/educatorsViewJson.js
