+++ app/assets/javascripts/feed/FeedView.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import FeedView from './FeedView';
 import {withDefaultNowContext} from '../testing/NowContainer';
-import homeFeedJson from '../../../../spec/javascripts/fixtures/home_feed_json';
+import homeFeedJson from '../testing/fixtures/home_feed_json';
 
 
 describe('FeedView', () => {
