+++ app/assets/javascripts/home/CheckStudentsWithHighAbsences.test.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import renderer from 'react-test-renderer';
 import fetchMock from 'fetch-mock/es5/client';
 import CheckStudentsWithHighAbsences, {CheckStudentsWithHighAbsencesView} from './CheckStudentsWithHighAbsences';
-import studentsWithHighAbsencesJson from '../../../../spec/javascripts/fixtures/home_students_with_high_absences_json';
+import studentsWithHighAbsencesJson from './home_students_with_high_absences_json';
 import {withDefaultNowContext} from '../testing/NowContainer';
 
 function renderIntoEl(element) {
