+++ app/assets/javascripts/homeroom_table/HomeroomTable.fixtures.js
@@ -1,3 +1,17 @@
+export function healey() {
+  return {
+    local_id: 'HEA',
+    school_type: 'ESMS'
+  };
+}
+
+export function shs() {
+  return {
+    local_id: 'SHS',
+    school_type: 'HS'
+  };
+}
+
 export const students = [
   {
     "id": 4,
