+++ app/assets/javascripts/school_courses/schoolCoursesJson.js
