+++ app/assets/javascripts/school_overview/CollapsableTable.js
