+++ app/assets/javascripts/school_overview/CollapsableTable.test.js
