+++ app/assets/javascripts/school_overview/FixedTable.js
