+++ app/assets/javascripts/school_overview/FixedTable.test.js
