+++ app/assets/javascripts/school_overview/SliceButtons.js
