+++ app/assets/javascripts/school_overview/SliceButtons.test.js
