+++ app/assets/javascripts/school_overview/SlicePanels.js
