+++ app/assets/javascripts/school_overview/SlicePanels.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import _ from 'lodash';
-import {serviceTypesIndex, eventNoteTypesIndex} from '../../../../spec/javascripts/fixtures/database_constants';
-import FixtureStudents from '../../../../spec/javascripts/fixtures/students';
+import {serviceTypesIndex, eventNoteTypesIndex} from '../testing/fixtures/databaseConstants';
+import FixtureStudents from './schoolOverviewStudents';
 import SlicePanels from './SlicePanels';
 
 function testProps(props = {}) {
