+++ app/assets/javascripts/school_overview/StudentsTable.js
