+++ app/assets/javascripts/school_overview/schoolOverviewStudents.fixture.js
@@ -1,0 +1,1 @@
+export default [{"id":1,"grade":"KF","hispanic_latino":null,"race":null,"free_reduced_lunch":null,"created_at":"2018-06-08T18:53:53.424Z","updated_at":"2018-06-08T18:53:53.424Z","homeroom_id":1,"first_name":"Garfield","last_name":"Skywalker","state_id":null,"home_language":null,"school_id":2,"registration_date":null,"local_id":"111111111","program_assigned":null,"sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":null,"limited_english_proficiency":null,"most_recent_mcas_math_growth":null,"most_recent_mcas_ela_growth":null,"most_recent_mcas_math_performance":null,"most_recent_mcas_ela_performance":null,"most_recent_mcas_math_scaled":null,"most_recent_mcas_ela_scaled":null,"most_recent_star_reading_percentile":null,"most_recent_star_math_percentile":null,"enrollment_status":"Active","date_of_birth":null,"risk_level":null,"gender":null,"house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":"HEA 003","event_notes":[],"active_services":[],"summer_services":[],"interventions":[]},{"id":6,"grade":"KF","hispanic_latino":true,"race":"Black","free_reduced_lunch":"Free Lunch","created_at":"2018-06-08T18:53:54.914Z","updated_at":"2018-06-08T18:54:48.313Z","homeroom_id":1,"first_name":"Winnie","last_name":"Mouse","state_id":null,"home_language":"English","school_id":2,"registration_date":null,"local_id":"000420","program_assigned":"2Way Spanish","sped_placement":null,"disability":"Emotional","sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP","most_recent_mcas_math_growth":46,"most_recent_mcas_ela_growth":71,"most_recent_mcas_math_performance":"Exceeding Expectations","most_recent_mcas_ela_performance":"Meeting Expectations","most_recent_mcas_math_scaled":567,"most_recent_mcas_ela_scaled":502,"most_recent_star_reading_percentile":76,"most_recent_star_math_percentile":91,"enrollment_status":"Active","date_of_birth":"2013-10-09T00:00:00.000Z","risk_level":0,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":"HEA 003","event_notes":[{"id":14,"student_id":6,"educator_id":1,"event_note_type_id":306,"recorded_at":"2010-10-01T00:00:00.000Z","is_restricted":false},{"id":15,"student_id":6,"educator_id":1,"event_note_type_id":305,"recorded_at":"2010-11-27T00:00:00.000Z","is_restricted":false},{"id":16,"student_id":6,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-01-15T00:00:00.000Z","is_restricted":false},{"id":17,"student_id":6,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-02-22T00:00:00.000Z","is_restricted":false},{"id":18,"student_id":6,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-04-16T00:00:00.000Z","is_restricted":false},{"id":19,"student_id":6,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-06-07T00:00:00.000Z","is_restricted":false},{"id":20,"student_id":6,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-07-07T00:00:00.000Z","is_restricted":false},{"id":21,"student_id":6,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-08-31T00:00:00.000Z","is_restricted":false},{"id":22,"student_id":6,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-10-14T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[{"id":1,"name":"51a Filing","intervention_type_id":45,"comment":"The student had a tough day.","goal":"The student will pass the assessment.","start_date":"October 19, 2010","start_date_timestamp":"2010-10-19","end_date":"November  2, 2010","educator_email":"rich@demo.studentinsights.org","educator_id":1},{"id":2,"name":"Behavior Plan","intervention_type_id":24,"comment":"This intervention has worked well.","goal":"The student will pass the assessment.","start_date":"December 16, 2010","start_date_timestamp":"2010-12-16","end_date":"December 30, 2010","educator_email":"rich@demo.studentinsights.org","educator_id":1},{"id":3,"name":"X Block Tutor","intervention_type_id":44,"comment":"Something happened.","goal":"Increase growth percentile for this academic subject.","start_date":"February  4, 2011","start_date_timestamp":"2011-02-04","end_date":"February 18, 2011","educator_email":"rich@demo.studentinsights.org","educator_id":1},{"id":4,"name":"51a Filing","intervention_type_id":45,"comment":"The student had a tough day.","goal":"Increase how often the student is engaged in positive behaviors.","start_date":"March 21, 2011","start_date_timestamp":"2011-03-21","end_date":"April  4, 2011","educator_email":"rich@demo.studentinsights.org","educator_id":1},{"id":5,"name":"Classroom Behavior Intervention","intervention_type_id":27,"comment":"The student had a tough day.","goal":"The student will pass the assessment.","start_date":"May 12, 2011","start_date_timestamp":"2011-05-12","end_date":"May 26, 2011","educator_email":"rich@demo.studentinsights.org","educator_id":1},{"id":6,"name":"Attendance Contract","intervention_type_id":22,"comment":"The student had a tough day.","goal":"Increase how often the student is engaged in positive behaviors.","start_date":"July  2, 2011","start_date_timestamp":"2011-07-02","end_date":"July 16, 2011","educator_email":"rich@demo.studentinsights.org","educator_id":1}]},{"id":7,"grade":"KF","hispanic_latino":true,"race":"Asian","free_reduced_lunch":"Reduced Lunch","created_at":"2018-06-08T18:53:55.422Z","updated_at":"2018-06-08T18:54:48.335Z","homeroom_id":1,"first_name":"Winnie","last_name":"Pan","state_id":null,"home_language":"Portuguese","school_id":2,"registration_date":null,"local_id":"000224","program_assigned":"2Way Spanish","sped_placement":null,"disability":"Specific LDs","sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Limited","most_recent_mcas_math_growth":81,"most_recent_mcas_ela_growth":78,"most_recent_mcas_math_performance":"Exceeding Expectations","most_recent_mcas_ela_performance":"A","most_recent_mcas_math_scaled":557,"most_recent_mcas_ela_scaled":247,"most_recent_star_reading_percentile":64,"most_recent_star_math_percentile":92,"enrollment_status":"Active","date_of_birth":"2014-05-16T00:00:00.000Z","risk_level":3,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":1,"absences_count":3,"tardies_count":1,"homeroom_name":"HEA 003","event_notes":[{"id":23,"student_id":7,"educator_id":1,"event_note_type_id":304,"recorded_at":"2010-10-21T00:00:00.000Z","is_restricted":false},{"id":24,"student_id":7,"educator_id":1,"event_note_type_id":305,"recorded_at":"2010-12-03T00:00:00.000Z","is_restricted":false},{"id":25,"student_id":7,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-01-03T00:00:00.000Z","is_restricted":false},{"id":26,"student_id":7,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-02-13T00:00:00.000Z","is_restricted":false},{"id":27,"student_id":7,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-04-08T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":9,"grade":"KF","hispanic_latino":true,"race":"White","free_reduced_lunch":"Free Lunch","created_at":"2018-06-08T18:53:56.403Z","updated_at":"2018-06-08T18:54:48.381Z","homeroom_id":1,"first_name":"Mowgli","last_name":"Duck","state_id":null,"home_language":"Portuguese","school_id":2,"registration_date":null,"local_id":"000593","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP-Transitioning","most_recent_mcas_math_growth":6,"most_recent_mcas_ela_growth":3,"most_recent_mcas_math_performance":"Partially Meeting","most_recent_mcas_ela_performance":"Not Meeting Expectations","most_recent_mcas_math_scaled":477,"most_recent_mcas_ela_scaled":490,"most_recent_star_reading_percentile":33,"most_recent_star_math_percentile":59,"enrollment_status":"Active","date_of_birth":"2013-12-07T00:00:00.000Z","risk_level":1,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":1,"tardies_count":2,"homeroom_name":"HEA 003","event_notes":[{"id":35,"student_id":9,"educator_id":1,"event_note_type_id":304,"recorded_at":"2010-10-02T00:00:00.000Z","is_restricted":false},{"id":36,"student_id":9,"educator_id":1,"event_note_type_id":306,"recorded_at":"2010-11-10T00:00:00.000Z","is_restricted":false},{"id":37,"student_id":9,"educator_id":1,"event_note_type_id":301,"recorded_at":"2010-12-12T00:00:00.000Z","is_restricted":false},{"id":38,"student_id":9,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-02-03T00:00:00.000Z","is_restricted":false},{"id":39,"student_id":9,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-03-27T00:00:00.000Z","is_restricted":false},{"id":40,"student_id":9,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-05-25T00:00:00.000Z","is_restricted":false},{"id":41,"student_id":9,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-07-04T00:00:00.000Z","is_restricted":false},{"id":42,"student_id":9,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-08-07T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":10,"grade":"KF","hispanic_latino":true,"race":"White","free_reduced_lunch":"Not Eligible","created_at":"2018-06-08T18:53:56.894Z","updated_at":"2018-06-08T18:54:48.406Z","homeroom_id":1,"first_name":"Mickey","last_name":"White","state_id":null,"home_language":"Spanish","school_id":2,"registration_date":null,"local_id":"000217","program_assigned":"Sp Ed","sped_placement":"Partial Inclusion","disability":"Specific LDs","sped_level_of_need":"Moderate","plan_504":"Not 504","limited_english_proficiency":"FLEP-Transitioning","most_recent_mcas_math_growth":40,"most_recent_mcas_ela_growth":39,"most_recent_mcas_math_performance":"Partially Meeting","most_recent_mcas_ela_performance":"Meeting Expectations","most_recent_mcas_math_scaled":423,"most_recent_mcas_ela_scaled":535,"most_recent_star_reading_percentile":19,"most_recent_star_math_percentile":96,"enrollment_status":"Active","date_of_birth":"2013-11-24T00:00:00.000Z","risk_level":2,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":3,"absences_count":0,"tardies_count":0,"homeroom_name":"HEA 003","event_notes":[{"id":43,"student_id":10,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-10-27T00:00:00.000Z","is_restricted":false},{"id":44,"student_id":10,"educator_id":1,"event_note_type_id":300,"recorded_at":"2010-11-27T00:00:00.000Z","is_restricted":false},{"id":45,"student_id":10,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-01-08T00:00:00.000Z","is_restricted":false},{"id":46,"student_id":10,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-03-08T00:00:00.000Z","is_restricted":false},{"id":47,"student_id":10,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-05-03T00:00:00.000Z","is_restricted":false},{"id":48,"student_id":10,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-06-04T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":11,"grade":"KF","hispanic_latino":false,"race":"Black","free_reduced_lunch":"Not Eligible","created_at":"2018-06-08T18:53:57.313Z","updated_at":"2018-06-08T18:54:48.431Z","homeroom_id":1,"first_name":"Pocahontas","last_name":"Disney","state_id":null,"home_language":"Portuguese","school_id":2,"registration_date":null,"local_id":"000219","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Limited","most_recent_mcas_math_growth":40,"most_recent_mcas_ela_growth":12,"most_recent_mcas_math_performance":"Not Meeting Expectations","most_recent_mcas_ela_performance":"Partially Meeting","most_recent_mcas_math_scaled":552,"most_recent_mcas_ela_scaled":468,"most_recent_star_reading_percentile":39,"most_recent_star_math_percentile":37,"enrollment_status":"Active","date_of_birth":"2014-08-11T00:00:00.000Z","risk_level":3,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":"HEA 003","event_notes":[{"id":49,"student_id":11,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-10-22T00:00:00.000Z","is_restricted":false},{"id":50,"student_id":11,"educator_id":1,"event_note_type_id":304,"recorded_at":"2010-12-14T00:00:00.000Z","is_restricted":false},{"id":51,"student_id":11,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-02-08T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":12,"grade":"KF","hispanic_latino":false,"race":"Asian","free_reduced_lunch":"Free Lunch","created_at":"2018-06-08T18:53:57.680Z","updated_at":"2018-06-08T18:54:48.457Z","homeroom_id":1,"first_name":"Aladdin","last_name":"White","state_id":null,"home_language":"Portuguese","school_id":2,"registration_date":null,"local_id":"000975","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Fluent","most_recent_mcas_math_growth":89,"most_recent_mcas_ela_growth":87,"most_recent_mcas_math_performance":"NI","most_recent_mcas_ela_performance":"Exceeding Expectations","most_recent_mcas_math_scaled":230,"most_recent_mcas_ela_scaled":466,"most_recent_star_reading_percentile":39,"most_recent_star_math_percentile":88,"enrollment_status":"Active","date_of_birth":"2014-02-21T00:00:00.000Z","risk_level":2,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":3,"tardies_count":0,"homeroom_name":"HEA 003","event_notes":[{"id":52,"student_id":12,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-10-13T00:00:00.000Z","is_restricted":false}],"active_services":[{"id":1,"student_id":12,"provided_by_educator_name":"Districtwide, Rich","service_type_id":509,"date_started":"2018-05-06T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":5,"recorded_at":"2018-05-06T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":2,"student_id":12,"provided_by_educator_name":null,"service_type_id":505,"date_started":"2018-03-17T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":11,"recorded_at":"2018-03-21T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":3,"student_id":12,"provided_by_educator_name":null,"service_type_id":510,"date_started":"2018-01-24T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":6,"recorded_at":"2018-01-27T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":4,"student_id":12,"provided_by_educator_name":"Districtwide, Rich","service_type_id":514,"date_started":"2017-12-17T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":11,"recorded_at":"2017-12-23T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"summer_services":[{"id":1,"student_id":12,"provided_by_educator_name":"Districtwide, Rich","service_type_id":509,"date_started":"2018-05-06T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":5,"recorded_at":"2018-05-06T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":3,"student_id":12,"provided_by_educator_name":null,"service_type_id":510,"date_started":"2018-01-24T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":6,"recorded_at":"2018-01-27T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"interventions":[]},{"id":13,"grade":"5","hispanic_latino":true,"race":"Asian","free_reduced_lunch":"Reduced Lunch","created_at":"2018-06-08T18:53:58.122Z","updated_at":"2018-06-08T18:54:48.483Z","homeroom_id":2,"first_name":"Mickey","last_name":"Mouse","state_id":null,"home_language":"Haitian-Creole","school_id":2,"registration_date":null,"local_id":"000263","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Fluent","most_recent_mcas_math_growth":46,"most_recent_mcas_ela_growth":34,"most_recent_mcas_math_performance":"NI","most_recent_mcas_ela_performance":"Meeting Expectations","most_recent_mcas_math_scaled":275,"most_recent_mcas_ela_scaled":412,"most_recent_star_reading_percentile":81,"most_recent_star_math_percentile":51,"enrollment_status":"Active","date_of_birth":"2008-09-04T00:00:00.000Z","risk_level":2,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":3,"tardies_count":1,"homeroom_name":"HEA 500","event_notes":[],"active_services":[],"summer_services":[],"interventions":[]},{"id":15,"grade":"5","hispanic_latino":false,"race":"White","free_reduced_lunch":null,"created_at":"2018-06-08T18:53:58.932Z","updated_at":"2018-06-08T18:54:48.531Z","homeroom_id":2,"first_name":"Pluto","last_name":"Poppins","state_id":null,"home_language":"Spanish","school_id":2,"registration_date":null,"local_id":"000226","program_assigned":"2Way English","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP","most_recent_mcas_math_growth":35,"most_recent_mcas_ela_growth":37,"most_recent_mcas_math_performance":"Meeting Expectations","most_recent_mcas_ela_performance":"Partially Meeting","most_recent_mcas_math_scaled":479,"most_recent_mcas_ela_scaled":596,"most_recent_star_reading_percentile":90,"most_recent_star_math_percentile":0,"enrollment_status":"Active","date_of_birth":"2009-03-13T00:00:00.000Z","risk_level":3,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":2,"tardies_count":2,"homeroom_name":"HEA 500","event_notes":[{"id":59,"student_id":15,"educator_id":1,"event_note_type_id":305,"recorded_at":"2010-10-12T00:00:00.000Z","is_restricted":false},{"id":60,"student_id":15,"educator_id":1,"event_note_type_id":306,"recorded_at":"2010-11-22T00:00:00.000Z","is_restricted":false},{"id":61,"student_id":15,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-01-14T00:00:00.000Z","is_restricted":false},{"id":62,"student_id":15,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-03-10T00:00:00.000Z","is_restricted":false},{"id":63,"student_id":15,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-05-02T00:00:00.000Z","is_restricted":false},{"id":64,"student_id":15,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-06-14T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":16,"grade":"5","hispanic_latino":false,"race":"Black","free_reduced_lunch":"Not Eligible","created_at":"2018-06-08T18:53:59.489Z","updated_at":"2018-06-08T18:54:48.554Z","homeroom_id":2,"first_name":"Elsa","last_name":"Pan","state_id":null,"home_language":"Spanish","school_id":2,"registration_date":null,"local_id":"000198","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP","most_recent_mcas_math_growth":11,"most_recent_mcas_ela_growth":34,"most_recent_mcas_math_performance":"Meeting Expectations","most_recent_mcas_ela_performance":"Partially Meeting","most_recent_mcas_math_scaled":548,"most_recent_mcas_ela_scaled":566,"most_recent_star_reading_percentile":0,"most_recent_star_math_percentile":89,"enrollment_status":"Active","date_of_birth":"2008-11-10T00:00:00.000Z","risk_level":3,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":1,"tardies_count":1,"homeroom_name":"HEA 500","event_notes":[{"id":65,"student_id":16,"educator_id":1,"event_note_type_id":305,"recorded_at":"2010-10-14T00:00:00.000Z","is_restricted":false},{"id":66,"student_id":16,"educator_id":1,"event_note_type_id":305,"recorded_at":"2010-12-02T00:00:00.000Z","is_restricted":false},{"id":67,"student_id":16,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-01-28T00:00:00.000Z","is_restricted":false},{"id":68,"student_id":16,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-03-15T00:00:00.000Z","is_restricted":false}],"active_services":[{"id":5,"student_id":16,"provided_by_educator_name":"Districtwide, Rich","service_type_id":504,"date_started":"2018-04-05T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":3,"recorded_at":"2018-04-10T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":6,"student_id":16,"provided_by_educator_name":"Districtwide, Rich","service_type_id":512,"date_started":"2018-02-14T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":13,"recorded_at":"2018-02-19T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"summer_services":[{"id":6,"student_id":16,"provided_by_educator_name":"Districtwide, Rich","service_type_id":512,"date_started":"2018-02-14T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":13,"recorded_at":"2018-02-19T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"interventions":[]},{"id":17,"grade":"5","hispanic_latino":false,"race":"White","free_reduced_lunch":null,"created_at":"2018-06-08T18:54:00.045Z","updated_at":"2018-06-08T18:54:48.576Z","homeroom_id":2,"first_name":"Winnie","last_name":"White","state_id":null,"home_language":"Portuguese","school_id":2,"registration_date":null,"local_id":"000381","program_assigned":"Reg Ed","sped_placement":null,"disability":"Emotional","sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP","most_recent_mcas_math_growth":82,"most_recent_mcas_ela_growth":38,"most_recent_mcas_math_performance":"Not Meeting Expectations","most_recent_mcas_ela_performance":"Meeting Expectations","most_recent_mcas_math_scaled":451,"most_recent_mcas_ela_scaled":462,"most_recent_star_reading_percentile":78,"most_recent_star_math_percentile":2,"enrollment_status":"Active","date_of_birth":"2009-06-20T00:00:00.000Z","risk_level":3,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":1,"tardies_count":0,"homeroom_name":"HEA 500","event_notes":[{"id":69,"student_id":17,"educator_id":1,"event_note_type_id":306,"recorded_at":"2010-10-23T00:00:00.000Z","is_restricted":false},{"id":70,"student_id":17,"educator_id":1,"event_note_type_id":305,"recorded_at":"2010-12-11T00:00:00.000Z","is_restricted":false},{"id":71,"student_id":17,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-02-03T00:00:00.000Z","is_restricted":false},{"id":72,"student_id":17,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-03-05T00:00:00.000Z","is_restricted":false},{"id":73,"student_id":17,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-04-15T00:00:00.000Z","is_restricted":false},{"id":74,"student_id":17,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-05-31T00:00:00.000Z","is_restricted":false},{"id":75,"student_id":17,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-07-30T00:00:00.000Z","is_restricted":false},{"id":76,"student_id":17,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-09-15T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":19,"grade":"5","hispanic_latino":false,"race":"Black","free_reduced_lunch":"Not Eligible","created_at":"2018-06-08T18:54:00.938Z","updated_at":"2018-06-08T18:54:48.628Z","homeroom_id":2,"first_name":"Rapunzel","last_name":"Kenobi","state_id":null,"home_language":"Haitian-Creole","school_id":2,"registration_date":null,"local_id":"000879","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Limited","most_recent_mcas_math_growth":85,"most_recent_mcas_ela_growth":65,"most_recent_mcas_math_performance":"Meeting Expectations","most_recent_mcas_ela_performance":"Not Meeting Expectations","most_recent_mcas_math_scaled":598,"most_recent_mcas_ela_scaled":448,"most_recent_star_reading_percentile":14,"most_recent_star_math_percentile":84,"enrollment_status":"Active","date_of_birth":"2009-02-07T00:00:00.000Z","risk_level":3,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":1,"tardies_count":2,"homeroom_name":"HEA 500","event_notes":[{"id":78,"student_id":19,"educator_id":1,"event_note_type_id":304,"recorded_at":"2010-10-18T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":21,"grade":"5","hispanic_latino":false,"race":"Asian","free_reduced_lunch":null,"created_at":"2018-06-08T18:54:01.709Z","updated_at":"2018-06-08T18:54:48.671Z","homeroom_id":2,"first_name":"Snow","last_name":"Disney","state_id":null,"home_language":"Haitian-Creole","school_id":2,"registration_date":null,"local_id":"000847","program_assigned":"2Way English","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP-Transitioning","most_recent_mcas_math_growth":85,"most_recent_mcas_ela_growth":19,"most_recent_mcas_math_performance":"Exceeding Expectations","most_recent_mcas_ela_performance":"Exceeding Expectations","most_recent_mcas_math_scaled":487,"most_recent_mcas_ela_scaled":414,"most_recent_star_reading_percentile":100,"most_recent_star_math_percentile":37,"enrollment_status":"Active","date_of_birth":"2009-04-08T00:00:00.000Z","risk_level":0,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":1,"homeroom_name":"HEA 500","event_notes":[],"active_services":[{"id":9,"student_id":21,"provided_by_educator_name":"Districtwide, Rich","service_type_id":510,"date_started":"2018-04-25T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":11,"recorded_at":"2018-04-30T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":10,"student_id":21,"provided_by_educator_name":null,"service_type_id":505,"date_started":"2018-02-26T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":4,"recorded_at":"2018-03-02T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":11,"student_id":21,"provided_by_educator_name":null,"service_type_id":504,"date_started":"2017-12-25T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":12,"recorded_at":"2018-01-01T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"summer_services":[{"id":9,"student_id":21,"provided_by_educator_name":"Districtwide, Rich","service_type_id":510,"date_started":"2018-04-25T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":11,"recorded_at":"2018-04-30T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"interventions":[]},{"id":4,"grade":"KF","hispanic_latino":true,"race":"Black","free_reduced_lunch":null,"created_at":"2018-06-08T18:53:53.882Z","updated_at":"2018-06-08T18:54:48.270Z","homeroom_id":1,"first_name":"Pocahontas","last_name":"Skywalker","state_id":null,"home_language":"English","school_id":2,"registration_date":null,"local_id":"000176","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Fluent","most_recent_mcas_math_growth":91,"most_recent_mcas_ela_growth":50,"most_recent_mcas_math_performance":"Meeting Expectations","most_recent_mcas_ela_performance":"A","most_recent_mcas_math_scaled":595,"most_recent_mcas_ela_scaled":259,"most_recent_star_reading_percentile":72,"most_recent_star_math_percentile":86,"enrollment_status":"Active","date_of_birth":"2014-06-05T00:00:00.000Z","risk_level":0,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":1,"absences_count":0,"tardies_count":1,"homeroom_name":"HEA 003","event_notes":[{"id":1,"student_id":4,"educator_id":1,"event_note_type_id":301,"recorded_at":"2010-10-03T00:00:00.000Z","is_restricted":false},{"id":2,"student_id":4,"educator_id":1,"event_note_type_id":301,"recorded_at":"2010-11-22T00:00:00.000Z","is_restricted":false},{"id":3,"student_id":4,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-01-04T00:00:00.000Z","is_restricted":false},{"id":4,"student_id":4,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-02-11T00:00:00.000Z","is_restricted":false},{"id":5,"student_id":4,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-03-24T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":58,"grade":"3","hispanic_latino":true,"race":"White","free_reduced_lunch":"Reduced Lunch","created_at":"2018-06-08T18:54:19.427Z","updated_at":"2018-06-08T18:54:49.530Z","homeroom_id":null,"first_name":"Elsa","last_name":"Pan","state_id":null,"home_language":"Spanish","school_id":2,"registration_date":null,"local_id":"000607","program_assigned":"2Way English","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Fluent","most_recent_mcas_math_growth":75,"most_recent_mcas_ela_growth":9,"most_recent_mcas_math_performance":"NI","most_recent_mcas_ela_performance":"Exceeding Expectations","most_recent_mcas_math_scaled":216,"most_recent_mcas_ela_scaled":463,"most_recent_star_reading_percentile":91,"most_recent_star_math_percentile":62,"enrollment_status":"Active","date_of_birth":"2011-01-27T00:00:00.000Z","risk_level":2,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":1,"absences_count":1,"tardies_count":4,"homeroom_name":null,"event_notes":[{"id":275,"student_id":58,"educator_id":1,"event_note_type_id":300,"recorded_at":"2010-10-12T00:00:00.000Z","is_restricted":false},{"id":276,"student_id":58,"educator_id":1,"event_note_type_id":301,"recorded_at":"2010-12-02T00:00:00.000Z","is_restricted":false},{"id":277,"student_id":58,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-01-01T00:00:00.000Z","is_restricted":false},{"id":278,"student_id":58,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-02-24T00:00:00.000Z","is_restricted":false},{"id":279,"student_id":58,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-04-09T00:00:00.000Z","is_restricted":false},{"id":280,"student_id":58,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-05-21T00:00:00.000Z","is_restricted":false}],"active_services":[{"id":38,"student_id":58,"provided_by_educator_name":null,"service_type_id":509,"date_started":"2018-05-04T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":7,"recorded_at":"2018-05-08T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":39,"student_id":58,"provided_by_educator_name":"Districtwide, Rich","service_type_id":503,"date_started":"2018-03-22T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":10,"recorded_at":"2018-03-27T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"summer_services":[{"id":38,"student_id":58,"provided_by_educator_name":null,"service_type_id":509,"date_started":"2018-05-04T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":7,"recorded_at":"2018-05-08T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"interventions":[]},{"id":59,"grade":"3","hispanic_latino":false,"race":"Black","free_reduced_lunch":"Not Eligible","created_at":"2018-06-08T18:54:19.897Z","updated_at":"2018-06-08T18:54:49.553Z","homeroom_id":null,"first_name":"Aladdin","last_name":"Duck","state_id":null,"home_language":"Portuguese","school_id":2,"registration_date":null,"local_id":"000261","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Limited","most_recent_mcas_math_growth":38,"most_recent_mcas_ela_growth":64,"most_recent_mcas_math_performance":"Exceeding Expectations","most_recent_mcas_ela_performance":null,"most_recent_mcas_math_scaled":450,"most_recent_mcas_ela_scaled":269,"most_recent_star_reading_percentile":13,"most_recent_star_math_percentile":0,"enrollment_status":"Active","date_of_birth":"2010-09-18T00:00:00.000Z","risk_level":3,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":4,"tardies_count":3,"homeroom_name":null,"event_notes":[{"id":282,"student_id":59,"educator_id":1,"event_note_type_id":305,"recorded_at":"2010-12-10T00:00:00.000Z","is_restricted":false},{"id":281,"student_id":59,"educator_id":1,"event_note_type_id":305,"recorded_at":"2010-10-25T00:00:00.000Z","is_restricted":false},{"id":283,"student_id":59,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-02-02T00:00:00.000Z","is_restricted":false},{"id":284,"student_id":59,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-03-24T00:00:00.000Z","is_restricted":false},{"id":285,"student_id":59,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-05-20T00:00:00.000Z","is_restricted":false},{"id":286,"student_id":59,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-07-05T00:00:00.000Z","is_restricted":false},{"id":287,"student_id":59,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-08-04T00:00:00.000Z","is_restricted":false},{"id":288,"student_id":59,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-09-28T00:00:00.000Z","is_restricted":false},{"id":289,"student_id":59,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-11-06T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":61,"grade":"3","hispanic_latino":true,"race":"Asian","free_reduced_lunch":null,"created_at":"2018-06-08T18:54:20.768Z","updated_at":"2018-06-08T18:54:49.602Z","homeroom_id":null,"first_name":"Elsa","last_name":"Kenobi","state_id":null,"home_language":"Spanish","school_id":2,"registration_date":null,"local_id":"000530","program_assigned":"2Way Spanish","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP-Transitioning","most_recent_mcas_math_growth":10,"most_recent_mcas_ela_growth":66,"most_recent_mcas_math_performance":"Meeting Expectations","most_recent_mcas_ela_performance":"Partially Meeting","most_recent_mcas_math_scaled":527,"most_recent_mcas_ela_scaled":492,"most_recent_star_reading_percentile":92,"most_recent_star_math_percentile":9,"enrollment_status":"Active","date_of_birth":"2011-04-03T00:00:00.000Z","risk_level":3,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":5,"tardies_count":2,"homeroom_name":null,"event_notes":[{"id":292,"student_id":61,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-10-03T00:00:00.000Z","is_restricted":false},{"id":293,"student_id":61,"educator_id":1,"event_note_type_id":305,"recorded_at":"2010-11-21T00:00:00.000Z","is_restricted":false},{"id":294,"student_id":61,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-01-02T00:00:00.000Z","is_restricted":false},{"id":295,"student_id":61,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-02-18T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":63,"grade":"3","hispanic_latino":true,"race":"Black","free_reduced_lunch":"Reduced Lunch","created_at":"2018-06-08T18:54:21.572Z","updated_at":"2018-06-08T18:54:49.650Z","homeroom_id":null,"first_name":"Pluto","last_name":"Skywalker","state_id":null,"home_language":"Spanish","school_id":2,"registration_date":null,"local_id":"000292","program_assigned":"Sp Ed","sped_placement":"Full Inclusion","disability":"Communication","sped_level_of_need":"High","plan_504":"Not 504","limited_english_proficiency":"Limited","most_recent_mcas_math_growth":62,"most_recent_mcas_ela_growth":79,"most_recent_mcas_math_performance":"NI","most_recent_mcas_ela_performance":"Not Meeting Expectations","most_recent_mcas_math_scaled":205,"most_recent_mcas_ela_scaled":531,"most_recent_star_reading_percentile":23,"most_recent_star_math_percentile":47,"enrollment_status":"Active","date_of_birth":"2011-06-11T00:00:00.000Z","risk_level":3,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":3,"tardies_count":3,"homeroom_name":null,"event_notes":[{"id":301,"student_id":63,"educator_id":1,"event_note_type_id":305,"recorded_at":"2010-10-07T00:00:00.000Z","is_restricted":false},{"id":302,"student_id":63,"educator_id":1,"event_note_type_id":301,"recorded_at":"2010-11-26T00:00:00.000Z","is_restricted":false},{"id":303,"student_id":63,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-01-10T00:00:00.000Z","is_restricted":false},{"id":304,"student_id":63,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-02-16T00:00:00.000Z","is_restricted":false},{"id":305,"student_id":63,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-03-28T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":64,"grade":"3","hispanic_latino":true,"race":"White","free_reduced_lunch":"Not Eligible","created_at":"2018-06-08T18:54:21.991Z","updated_at":"2018-06-08T18:54:49.673Z","homeroom_id":null,"first_name":"Winnie","last_name":"White","state_id":null,"home_language":"Spanish","school_id":2,"registration_date":null,"local_id":"000112","program_assigned":"2Way Spanish","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP-Transitioning","most_recent_mcas_math_growth":75,"most_recent_mcas_ela_growth":0,"most_recent_mcas_math_performance":"Not Meeting Expectations","most_recent_mcas_ela_performance":"Partially Meeting","most_recent_mcas_math_scaled":406,"most_recent_mcas_ela_scaled":510,"most_recent_star_reading_percentile":94,"most_recent_star_math_percentile":90,"enrollment_status":"Active","date_of_birth":"2011-04-19T00:00:00.000Z","risk_level":0,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":1,"absences_count":3,"tardies_count":7,"homeroom_name":null,"event_notes":[],"active_services":[],"summer_services":[],"interventions":[]},{"id":66,"grade":"3","hispanic_latino":true,"race":"White","free_reduced_lunch":"Not Eligible","created_at":"2018-06-08T18:54:22.608Z","updated_at":"2018-06-08T18:54:49.721Z","homeroom_id":null,"first_name":"Aladdin","last_name":"Disney","state_id":null,"home_language":"Spanish","school_id":2,"registration_date":null,"local_id":"000533","program_assigned":"2Way Spanish","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Limited","most_recent_mcas_math_growth":67,"most_recent_mcas_ela_growth":54,"most_recent_mcas_math_performance":"NI","most_recent_mcas_ela_performance":"Not Meeting Expectations","most_recent_mcas_math_scaled":238,"most_recent_mcas_ela_scaled":565,"most_recent_star_reading_percentile":7,"most_recent_star_math_percentile":77,"enrollment_status":"Active","date_of_birth":"2011-08-11T00:00:00.000Z","risk_level":3,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":null,"event_notes":[{"id":309,"student_id":66,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-10-19T00:00:00.000Z","is_restricted":false},{"id":310,"student_id":66,"educator_id":1,"event_note_type_id":305,"recorded_at":"2010-11-29T00:00:00.000Z","is_restricted":false},{"id":311,"student_id":66,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-01-02T00:00:00.000Z","is_restricted":false},{"id":312,"student_id":66,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-02-21T00:00:00.000Z","is_restricted":false},{"id":313,"student_id":66,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-03-28T00:00:00.000Z","is_restricted":false},{"id":314,"student_id":66,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-04-29T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":65,"grade":"3","hispanic_latino":false,"race":"Asian","free_reduced_lunch":null,"created_at":"2018-06-08T18:54:22.554Z","updated_at":"2018-06-08T18:54:46.859Z","homeroom_id":null,"first_name":"Pluto","last_name":"Pan","state_id":null,"home_language":"English","school_id":2,"registration_date":null,"local_id":"000797","program_assigned":"2Way English","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Limited","most_recent_mcas_math_growth":null,"most_recent_mcas_ela_growth":null,"most_recent_mcas_math_performance":null,"most_recent_mcas_ela_performance":null,"most_recent_mcas_math_scaled":null,"most_recent_mcas_ela_scaled":null,"most_recent_star_reading_percentile":null,"most_recent_star_math_percentile":null,"enrollment_status":"Active","date_of_birth":"2010-12-13T00:00:00.000Z","risk_level":3,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":null,"event_notes":[{"id":306,"student_id":65,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-10-14T00:00:00.000Z","is_restricted":false},{"id":307,"student_id":65,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-11-24T00:00:00.000Z","is_restricted":false},{"id":308,"student_id":65,"educator_id":1,"event_note_type_id":304,"recorded_at":"2010-12-31T00:00:00.000Z","is_restricted":false}],"active_services":[{"id":40,"student_id":65,"provided_by_educator_name":"Districtwide, Rich","service_type_id":503,"date_started":"2018-04-16T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":2,"recorded_at":"2018-04-17T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":41,"student_id":65,"provided_by_educator_name":null,"service_type_id":509,"date_started":"2018-03-01T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":3,"recorded_at":"2018-03-01T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":42,"student_id":65,"provided_by_educator_name":null,"service_type_id":514,"date_started":"2017-12-31T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":9,"recorded_at":"2017-12-31T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"summer_services":[{"id":41,"student_id":65,"provided_by_educator_name":null,"service_type_id":509,"date_started":"2018-03-01T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":3,"recorded_at":"2018-03-01T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"interventions":[]},{"id":8,"grade":"KF","hispanic_latino":false,"race":"Asian","free_reduced_lunch":"Free Lunch","created_at":"2018-06-08T18:53:55.930Z","updated_at":"2018-06-08T18:54:48.359Z","homeroom_id":1,"first_name":"Donald","last_name":"Kenobi","state_id":null,"home_language":"Spanish","school_id":2,"registration_date":null,"local_id":"000152","program_assigned":"2Way Spanish","sped_placement":null,"disability":"Communication","sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP-Transitioning","most_recent_mcas_math_growth":63,"most_recent_mcas_ela_growth":96,"most_recent_mcas_math_performance":"Meeting Expectations","most_recent_mcas_ela_performance":"Exceeding Expectations","most_recent_mcas_math_scaled":577,"most_recent_mcas_ela_scaled":578,"most_recent_star_reading_percentile":24,"most_recent_star_math_percentile":31,"enrollment_status":"Active","date_of_birth":"2013-09-29T00:00:00.000Z","risk_level":2,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":2,"tardies_count":0,"homeroom_name":"HEA 003","event_notes":[{"id":28,"student_id":8,"educator_id":1,"event_note_type_id":304,"recorded_at":"2010-10-21T00:00:00.000Z","is_restricted":false},{"id":29,"student_id":8,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-12-09T00:00:00.000Z","is_restricted":false},{"id":30,"student_id":8,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-01-29T00:00:00.000Z","is_restricted":false},{"id":31,"student_id":8,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-03-18T00:00:00.000Z","is_restricted":false},{"id":32,"student_id":8,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-04-26T00:00:00.000Z","is_restricted":false},{"id":33,"student_id":8,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-05-30T00:00:00.000Z","is_restricted":false},{"id":34,"student_id":8,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-07-13T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":14,"grade":"5","hispanic_latino":false,"race":"Asian","free_reduced_lunch":"Reduced Lunch","created_at":"2018-06-08T18:53:58.515Z","updated_at":"2018-06-08T18:54:48.509Z","homeroom_id":2,"first_name":"Chip","last_name":"Kenobi","state_id":null,"home_language":"Haitian-Creole","school_id":2,"registration_date":null,"local_id":"000385","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP","most_recent_mcas_math_growth":11,"most_recent_mcas_ela_growth":67,"most_recent_mcas_math_performance":"Exceeding Expectations","most_recent_mcas_ela_performance":"Not Meeting Expectations","most_recent_mcas_math_scaled":445,"most_recent_mcas_ela_scaled":515,"most_recent_star_reading_percentile":35,"most_recent_star_math_percentile":44,"enrollment_status":"Active","date_of_birth":"2008-12-10T00:00:00.000Z","risk_level":1,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":2,"absences_count":4,"tardies_count":1,"homeroom_name":"HEA 500","event_notes":[{"id":53,"student_id":14,"educator_id":1,"event_note_type_id":305,"recorded_at":"2010-10-09T00:00:00.000Z","is_restricted":false},{"id":54,"student_id":14,"educator_id":1,"event_note_type_id":300,"recorded_at":"2010-11-16T00:00:00.000Z","is_restricted":false},{"id":55,"student_id":14,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-01-15T00:00:00.000Z","is_restricted":false},{"id":56,"student_id":14,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-02-25T00:00:00.000Z","is_restricted":false},{"id":57,"student_id":14,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-04-09T00:00:00.000Z","is_restricted":false},{"id":58,"student_id":14,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-06-04T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":18,"grade":"5","hispanic_latino":true,"race":"Black","free_reduced_lunch":"Reduced Lunch","created_at":"2018-06-08T18:54:00.488Z","updated_at":"2018-06-08T18:54:48.606Z","homeroom_id":2,"first_name":"Winnie","last_name":"Skywalker","state_id":null,"home_language":"Portuguese","school_id":2,"registration_date":null,"local_id":"000678","program_assigned":"2Way English","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP-Transitioning","most_recent_mcas_math_growth":89,"most_recent_mcas_ela_growth":49,"most_recent_mcas_math_performance":"Meeting Expectations","most_recent_mcas_ela_performance":"Exceeding Expectations","most_recent_mcas_math_scaled":459,"most_recent_mcas_ela_scaled":566,"most_recent_star_reading_percentile":19,"most_recent_star_math_percentile":85,"enrollment_status":"Active","date_of_birth":"2008-11-04T00:00:00.000Z","risk_level":2,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":1,"tardies_count":2,"homeroom_name":"HEA 500","event_notes":[{"id":77,"student_id":18,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-10-12T00:00:00.000Z","is_restricted":false}],"active_services":[{"id":7,"student_id":18,"provided_by_educator_name":null,"service_type_id":509,"date_started":"2018-04-07T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":999999,"recorded_at":"2018-04-10T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":8,"student_id":18,"provided_by_educator_name":"Districtwide, Rich","service_type_id":505,"date_started":"2018-02-23T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":11,"recorded_at":"2018-02-26T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"summer_services":[{"id":7,"student_id":18,"provided_by_educator_name":null,"service_type_id":509,"date_started":"2018-04-07T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":999999,"recorded_at":"2018-04-10T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"interventions":[]},{"id":62,"grade":"3","hispanic_latino":false,"race":"Asian","free_reduced_lunch":null,"created_at":"2018-06-08T18:54:21.219Z","updated_at":"2018-06-08T18:54:49.627Z","homeroom_id":null,"first_name":"Elsa","last_name":"Duck","state_id":null,"home_language":"Spanish","school_id":2,"registration_date":null,"local_id":"000472","program_assigned":"Sp Ed","sped_placement":"Full Inclusion","disability":"Specific LDs","sped_level_of_need":"High","plan_504":"Not 504","limited_english_proficiency":"FLEP-Transitioning","most_recent_mcas_math_growth":20,"most_recent_mcas_ela_growth":48,"most_recent_mcas_math_performance":"Not Meeting Expectations","most_recent_mcas_ela_performance":"Exceeding Expectations","most_recent_mcas_math_scaled":458,"most_recent_mcas_ela_scaled":507,"most_recent_star_reading_percentile":71,"most_recent_star_math_percentile":57,"enrollment_status":"Active","date_of_birth":"2010-10-31T00:00:00.000Z","risk_level":1,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":null,"event_notes":[{"id":296,"student_id":62,"educator_id":1,"event_note_type_id":304,"recorded_at":"2010-10-01T00:00:00.000Z","is_restricted":false},{"id":297,"student_id":62,"educator_id":1,"event_note_type_id":306,"recorded_at":"2010-11-08T00:00:00.000Z","is_restricted":false},{"id":298,"student_id":62,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-01-05T00:00:00.000Z","is_restricted":false},{"id":299,"student_id":62,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-02-11T00:00:00.000Z","is_restricted":false},{"id":300,"student_id":62,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-03-21T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]}];
