+++ app/assets/javascripts/sessionTimeoutWarning.js
@@ -1,23 +1,20 @@
+function count(sessionTimeoutInSeconds) {
+  setTimeout(show, sessionTimeoutInSeconds * 1000);
+}
+
+function show() {
+  $('#renew-session').slideDown();
+}
+
 export default function sessionTimeoutWarning(Env) {
-  const SessionTimeoutWarning = function() {};
-
-  SessionTimeoutWarning.prototype.count = () => {
-    setTimeout(this.show, Env.sessionTimeoutInSeconds * 1000);
-  };
-
-  SessionTimeoutWarning.prototype.show = () => {
-    $('#renew-session').slideDown();
-  };
-
-  const warning = new SessionTimeoutWarning;
-  warning.count();
+  count(Env.sessionTimeoutInSeconds);
 
   $("#renew-sesion-link").click(() => {
     $.ajax({
       url: '/educators/reset',
       success() {
         $('#renew-session').slideUp();
-        warning.count();   // Resent timeout count
+        count(Env.sessionTimeoutInSeconds);   // Resent timeout count
       }
     });
   });
