+++ app/assets/javascripts/student_profile/fixtures.js
@@ -1,9 +1,8 @@
 import moment from 'moment';
 import {
-
   serviceTypesIndex,
   eventNoteTypesIndex
-} from '../../../../spec/javascripts/fixtures/database_constants';
+} from '../testing/fixtures/databaseConstants';
 
 // assuming static time for specs
 export const nowMoment = moment('2016-02-11T10:15:00');
