+++ app/assets/javascripts/testing/fixtures/README.md
@@ -1,0 +1,3 @@
+Put fixtures next to code and tests.
+
+Only put them here if you need to share them across different parts of the app.
