+++ app/assets/javascripts/testing/fixtures/databaseConstants.js
