--- spec/javascripts/fixtures/homeroom_students.js
@@ -1,822 +1,0 @@
-export default [
-  {
-    "id": 4,
-    "grade": "5",
-    "hispanic_latino": true,
-    "race": "White",
-    "free_reduced_lunch": "Free Lunch",
-    "created_at": "2018-01-10T13:42:06.268Z",
-    "updated_at": "2018-01-10T13:43:00.637Z",
-    "homeroom_id": 2,
-    "first_name": "Snow",
-    "last_name": "Skywalker",
-    "state_id": null,
-    "home_language": "English",
-    "school_id": 2,
-    "registration_date": null,
-    "local_id": "000209",
-    "program_assigned": "2Way Spanish",
-    "sped_placement": null,
-    "disability": null,
-    "sped_level_of_need": null,
-    "plan_504": "Not 504",
-    "limited_english_proficiency": "Fluent",
-    "most_recent_mcas_math_growth": 74,
-    "most_recent_mcas_ela_growth": 40,
-    "most_recent_mcas_math_performance": "Exceeding Expectations",
-    "most_recent_mcas_ela_performance": "Meeting Expectations",
-    "most_recent_mcas_math_scaled": 497,
-    "most_recent_mcas_ela_scaled": 499,
-    "most_recent_star_reading_percentile": 74,
-    "most_recent_star_math_percentile": 95,
-    "enrollment_status": "Active",
-    "date_of_birth": "2009-05-03T00:00:00.000Z",
-    "risk_level": 0,
-    "gender": "F",
-    "house": null,
-    "counselor": null,
-    "discipline_incidents_count": 0,
-    "absences_count": 2,
-    "tardies_count": 0,
-    "homeroom_name": "HEA 500",
-    "event_notes_without_restricted": [
-      {
-        "id": 3,
-        "student_id": 4,
-        "educator_id": 1,
-        "event_note_type_id": 300,
-        "text": "Meredith is setting up a family meeting to discuss absences and tardies, and to follow up on concerns about supporting morning routines at home.",
-        "recorded_at": "2010-10-31T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:06.335Z",
-        "updated_at": "2018-01-10T13:42:06.335Z",
-        "is_restricted": false
-      },
-      {
-        "id": 4,
-        "student_id": 4,
-        "educator_id": 1,
-        "event_note_type_id": 306,
-        "text": "Engagement issues in ELA, sometimes shutting down altogether and refusing to engage in academic work.  Other times this is more disruptive, lashing out and yelling at other students and interfering with their learning.",
-        "recorded_at": "2010-12-05T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:06.340Z",
-        "updated_at": "2018-01-10T13:42:06.340Z",
-        "is_restricted": false
-      },
-      {
-        "id": 5,
-        "student_id": 4,
-        "educator_id": 1,
-        "event_note_type_id": 306,
-        "text": "Lila will look into a referral to a big sister program.",
-        "recorded_at": "2011-01-08T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:06.345Z",
-        "updated_at": "2018-01-10T13:42:06.345Z",
-        "is_restricted": false
-      }
-    ],
-    "interventions": [],
-    "sped_data": {
-      "sped_level": "—",
-      "sped_tooltip_message": null,
-      "sped_bubble_class": "sped"
-    },
-    "student_risk_level": {
-      "id": 4,
-      "student_id": 4,
-      "level": 0,
-      "created_at": "2018-01-10T13:42:56.153Z",
-      "updated_at": "2018-01-10T13:42:56.184Z",
-      "mcas_math_risk_level": null,
-      "star_math_risk_level": 0,
-      "mcas_ela_risk_level": null,
-      "star_reading_risk_level": 1,
-      "limited_english_proficiency_risk_level": null,
-      "explanation": {
-        "intro": "Snow is at Risk 0 because:",
-        "reasons": [
-          "Snow's STAR Math performance is above 85."
-        ]
-      }
-    }
-  },
-  {
-    "id": 7,
-    "grade": "5",
-    "hispanic_latino": false,
-    "race": "White",
-    "free_reduced_lunch": "Not Eligible",
-    "created_at": "2018-01-10T13:42:07.457Z",
-    "updated_at": "2018-01-10T13:43:00.710Z",
-    "homeroom_id": 2,
-    "first_name": "Chip",
-    "last_name": "Pan",
-    "state_id": null,
-    "home_language": "English",
-    "school_id": 2,
-    "registration_date": null,
-    "local_id": "000991",
-    "program_assigned": "Reg Ed",
-    "sped_placement": null,
-    "disability": null,
-    "sped_level_of_need": null,
-    "plan_504": "504",
-    "limited_english_proficiency": "FLEP",
-    "most_recent_mcas_math_growth": 76,
-    "most_recent_mcas_ela_growth": 42,
-    "most_recent_mcas_math_performance": "Partially Meeting",
-    "most_recent_mcas_ela_performance": "Meeting Expectations",
-    "most_recent_mcas_math_scaled": 526,
-    "most_recent_mcas_ela_scaled": 512,
-    "most_recent_star_reading_percentile": 90,
-    "most_recent_star_math_percentile": 44,
-    "enrollment_status": "Active",
-    "date_of_birth": "2009-03-12T00:00:00.000Z",
-    "risk_level": 0,
-    "gender": "M",
-    "house": null,
-    "counselor": null,
-    "discipline_incidents_count": 0,
-    "absences_count": 1,
-    "tardies_count": 0,
-    "homeroom_name": "HEA 500",
-    "event_notes_without_restricted": [
-      {
-        "id": 18,
-        "student_id": 7,
-        "educator_id": 1,
-        "event_note_type_id": 306,
-        "text": "Philis is meeting with parent next week and will present an attendance contract.",
-        "recorded_at": "2010-10-30T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.505Z",
-        "updated_at": "2018-01-10T13:42:07.505Z",
-        "is_restricted": false
-      },
-      {
-        "id": 19,
-        "student_id": 7,
-        "educator_id": 1,
-        "event_note_type_id": 305,
-        "text": "Engagement issues in ELA, sometimes shutting down altogether and refusing to engage in academic work.  Other times this is more disruptive, lashing out and yelling at other students and interfering with their learning.",
-        "recorded_at": "2010-12-21T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.509Z",
-        "updated_at": "2018-01-10T13:42:07.509Z",
-        "is_restricted": false
-      },
-      {
-        "id": 20,
-        "student_id": 7,
-        "educator_id": 1,
-        "event_note_type_id": 300,
-        "text": "Engagement issues in ELA, sometimes shutting down altogether and refusing to engage in academic work.  Other times this is more disruptive, lashing out and yelling at other students and interfering with their learning.",
-        "recorded_at": "2011-02-16T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.514Z",
-        "updated_at": "2018-01-10T13:42:07.514Z",
-        "is_restricted": false
-      },
-      {
-        "id": 21,
-        "student_id": 7,
-        "educator_id": 1,
-        "event_note_type_id": 304,
-        "text": "Lila will look into a referral to a big sister program.",
-        "recorded_at": "2011-04-16T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.519Z",
-        "updated_at": "2018-01-10T13:42:07.519Z",
-        "is_restricted": false
-      },
-      {
-        "id": 22,
-        "student_id": 7,
-        "educator_id": 1,
-        "event_note_type_id": 305,
-        "text": "Not engaged in academic work and misbehaving in class.  There have been meetings between the teacher and support staff, and Sam has talked with Chip twice this week.  Philis following up with outside providers as well.",
-        "recorded_at": "2011-05-26T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.523Z",
-        "updated_at": "2018-01-10T13:42:07.523Z",
-        "is_restricted": false
-      },
-      {
-        "id": 23,
-        "student_id": 7,
-        "educator_id": 1,
-        "event_note_type_id": 302,
-        "text": "Lila will look into a referral to a big sister program.",
-        "recorded_at": "2011-06-25T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.528Z",
-        "updated_at": "2018-01-10T13:42:07.528Z",
-        "is_restricted": false
-      },
-      {
-        "id": 24,
-        "student_id": 7,
-        "educator_id": 1,
-        "event_note_type_id": 302,
-        "text": "Philis is meeting with parent next week and will present an attendance contract.",
-        "recorded_at": "2011-08-19T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.532Z",
-        "updated_at": "2018-01-10T13:42:07.532Z",
-        "is_restricted": false
-      }
-    ],
-    "interventions": [],
-    "sped_data": {
-      "sped_level": "—",
-      "sped_tooltip_message": null,
-      "sped_bubble_class": "sped"
-    },
-    "student_risk_level": {
-      "id": 7,
-      "student_id": 7,
-      "level": 0,
-      "created_at": "2018-01-10T13:42:56.306Z",
-      "updated_at": "2018-01-10T13:42:56.332Z",
-      "mcas_math_risk_level": null,
-      "star_math_risk_level": 1,
-      "mcas_ela_risk_level": null,
-      "star_reading_risk_level": 0,
-      "limited_english_proficiency_risk_level": null,
-      "explanation": {
-        "intro": "Chip is at Risk 0 because:",
-        "reasons": [
-          "Chip's STAR Reading performance is above 85."
-        ]
-      }
-    }
-  },
-  {
-    "id": 9,
-    "grade": "5",
-    "hispanic_latino": false,
-    "race": "Black",
-    "free_reduced_lunch": "Free Lunch",
-    "created_at": "2018-01-10T13:42:08.285Z",
-    "updated_at": "2018-01-10T13:43:00.758Z",
-    "homeroom_id": 2,
-    "first_name": "Pluto",
-    "last_name": "Poppins",
-    "state_id": null,
-    "home_language": "Haitian-Creole",
-    "school_id": 2,
-    "registration_date": null,
-    "local_id": "000612",
-    "program_assigned": "2Way English",
-    "sped_placement": null,
-    "disability": null,
-    "sped_level_of_need": null,
-    "plan_504": "Not 504",
-    "limited_english_proficiency": "Fluent",
-    "most_recent_mcas_math_growth": 95,
-    "most_recent_mcas_ela_growth": 32,
-    "most_recent_mcas_math_performance": "Not Meeting Expectations",
-    "most_recent_mcas_ela_performance": "Meeting Expectations",
-    "most_recent_mcas_math_scaled": 448,
-    "most_recent_mcas_ela_scaled": 451,
-    "most_recent_star_reading_percentile": 54,
-    "most_recent_star_math_percentile": 100,
-    "enrollment_status": "Active",
-    "date_of_birth": "2009-07-02T00:00:00.000Z",
-    "risk_level": 0,
-    "gender": "M",
-    "house": null,
-    "counselor": null,
-    "discipline_incidents_count": 0,
-    "absences_count": 2,
-    "tardies_count": 1,
-    "homeroom_name": "HEA 500",
-    "event_notes_without_restricted": [
-      {
-        "id": 30,
-        "student_id": 9,
-        "educator_id": 1,
-        "event_note_type_id": 300,
-        "text": "Meredith is setting up a family meeting to discuss absences and tardies, and to follow up on concerns about supporting morning routines at home.",
-        "recorded_at": "2010-10-02T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.348Z",
-        "updated_at": "2018-01-10T13:42:08.348Z",
-        "is_restricted": false
-      },
-      {
-        "id": 31,
-        "student_id": 9,
-        "educator_id": 1,
-        "event_note_type_id": 304,
-        "text": "Not engaged in academic work and misbehaving in class.  There have been meetings between the teacher and support staff, and Sam has talked with Pluto twice this week.  Philis following up with outside providers as well.",
-        "recorded_at": "2010-11-26T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.352Z",
-        "updated_at": "2018-01-10T13:42:08.352Z",
-        "is_restricted": false
-      },
-      {
-        "id": 32,
-        "student_id": 9,
-        "educator_id": 1,
-        "event_note_type_id": 301,
-        "text": "Meetings with staff about lack of academic progress.",
-        "recorded_at": "2011-01-19T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.357Z",
-        "updated_at": "2018-01-10T13:42:08.357Z",
-        "is_restricted": false
-      },
-      {
-        "id": 33,
-        "student_id": 9,
-        "educator_id": 1,
-        "event_note_type_id": 301,
-        "text": "Lila will look into a referral to a big sister program.",
-        "recorded_at": "2011-03-08T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.362Z",
-        "updated_at": "2018-01-10T13:42:08.362Z",
-        "is_restricted": false
-      },
-      {
-        "id": 34,
-        "student_id": 9,
-        "educator_id": 1,
-        "event_note_type_id": 302,
-        "text": "Meetings with staff about lack of academic progress.",
-        "recorded_at": "2011-04-15T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.367Z",
-        "updated_at": "2018-01-10T13:42:08.367Z",
-        "is_restricted": false
-      },
-      {
-        "id": 35,
-        "student_id": 9,
-        "educator_id": 1,
-        "event_note_type_id": 301,
-        "text": "Pluto will go through the special ed process for placement",
-        "recorded_at": "2011-06-05T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.372Z",
-        "updated_at": "2018-01-10T13:42:08.372Z",
-        "is_restricted": false
-      },
-      {
-        "id": 36,
-        "student_id": 9,
-        "educator_id": 1,
-        "event_note_type_id": 305,
-        "text": "Philis is meeting with parent next week and will present an attendance contract.",
-        "recorded_at": "2011-07-14T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.376Z",
-        "updated_at": "2018-01-10T13:42:08.376Z",
-        "is_restricted": false
-      },
-      {
-        "id": 37,
-        "student_id": 9,
-        "educator_id": 1,
-        "event_note_type_id": 300,
-        "text": "Engagement issues in ELA, sometimes shutting down altogether and refusing to engage in academic work.  Other times this is more disruptive, lashing out and yelling at other students and interfering with their learning.",
-        "recorded_at": "2011-09-02T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.381Z",
-        "updated_at": "2018-01-10T13:42:08.381Z",
-        "is_restricted": false
-      }
-    ],
-    "interventions": [],
-    "sped_data": {
-      "sped_level": "—",
-      "sped_tooltip_message": null,
-      "sped_bubble_class": "sped"
-    },
-    "student_risk_level": {
-      "id": 9,
-      "student_id": 9,
-      "level": 0,
-      "created_at": "2018-01-10T13:42:56.394Z",
-      "updated_at": "2018-01-10T13:42:56.421Z",
-      "mcas_math_risk_level": null,
-      "star_math_risk_level": 0,
-      "mcas_ela_risk_level": null,
-      "star_reading_risk_level": 1,
-      "limited_english_proficiency_risk_level": null,
-      "explanation": {
-        "intro": "Pluto is at Risk 0 because:",
-        "reasons": [
-          "Pluto's STAR Math performance is above 85."
-        ]
-      }
-    }
-  },
-  {
-    "id": 3,
-    "grade": "5",
-    "hispanic_latino": true,
-    "race": "White",
-    "free_reduced_lunch": "Free Lunch",
-    "created_at": "2018-01-10T13:42:05.749Z",
-    "updated_at": "2018-01-10T13:43:00.612Z",
-    "homeroom_id": 2,
-    "first_name": "Aladdin",
-    "last_name": "Kenobi",
-    "state_id": null,
-    "home_language": "Haitian-Creole",
-    "school_id": 2,
-    "registration_date": null,
-    "local_id": "00085",
-    "program_assigned": "2Way Spanish",
-    "sped_placement": null,
-    "disability": null,
-    "sped_level_of_need": null,
-    "plan_504": "Not 504",
-    "limited_english_proficiency": "FLEP",
-    "most_recent_mcas_math_growth": 33,
-    "most_recent_mcas_ela_growth": 2,
-    "most_recent_mcas_math_performance": "P",
-    "most_recent_mcas_ela_performance": "Meeting Expectations",
-    "most_recent_mcas_math_scaled": 255,
-    "most_recent_mcas_ela_scaled": 549,
-    "most_recent_star_reading_percentile": 76,
-    "most_recent_star_math_percentile": 2,
-    "enrollment_status": "Active",
-    "date_of_birth": "2009-08-13T00:00:00.000Z",
-    "risk_level": 3,
-    "gender": "F",
-    "house": null,
-    "counselor": null,
-    "discipline_incidents_count": 0,
-    "absences_count": 0,
-    "tardies_count": 0,
-    "homeroom_name": "HEA 500",
-    "event_notes_without_restricted": [
-      {
-        "id": 1,
-        "student_id": 3,
-        "educator_id": 1,
-        "event_note_type_id": 302,
-        "text": "Philis is meeting with parent next week and will present an attendance contract.",
-        "recorded_at": "2010-10-15T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:05.802Z",
-        "updated_at": "2018-01-10T13:42:05.802Z",
-        "is_restricted": false
-      },
-      {
-        "id": 2,
-        "student_id": 3,
-        "educator_id": 1,
-        "event_note_type_id": 304,
-        "text": "Meredith is setting up a family meeting to discuss absences and tardies, and to follow up on concerns about supporting morning routines at home.",
-        "recorded_at": "2010-12-10T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:05.810Z",
-        "updated_at": "2018-01-10T13:42:05.810Z",
-        "is_restricted": false
-      }
-    ],
-    "interventions": [],
-    "sped_data": {
-      "sped_level": "—",
-      "sped_tooltip_message": null,
-      "sped_bubble_class": "sped"
-    },
-    "student_risk_level": {
-      "id": 3,
-      "student_id": 3,
-      "level": 3,
-      "created_at": "2018-01-10T13:42:56.106Z",
-      "updated_at": "2018-01-10T13:42:56.139Z",
-      "mcas_math_risk_level": 1,
-      "star_math_risk_level": 3,
-      "mcas_ela_risk_level": null,
-      "star_reading_risk_level": 1,
-      "limited_english_proficiency_risk_level": null,
-      "explanation": {
-        "intro": "Aladdin is at Risk 3 because:",
-        "reasons": [
-          "Aladdin's STAR Math performance is in the warning range (below 10)."
-        ]
-      }
-    }
-  },
-  {
-    "id": 6,
-    "grade": "5",
-    "hispanic_latino": false,
-    "race": "Asian",
-    "free_reduced_lunch": "Free Lunch",
-    "created_at": "2018-01-10T13:42:07.064Z",
-    "updated_at": "2018-01-10T13:43:00.686Z",
-    "homeroom_id": 2,
-    "first_name": "Snow",
-    "last_name": "Kenobi",
-    "state_id": null,
-    "home_language": "English",
-    "school_id": 2,
-    "registration_date": null,
-    "local_id": "000407",
-    "program_assigned": "2Way English",
-    "sped_placement": null,
-    "disability": null,
-    "sped_level_of_need": null,
-    "plan_504": "Not 504",
-    "limited_english_proficiency": "FLEP-Transitioning",
-    "most_recent_mcas_math_growth": 0,
-    "most_recent_mcas_ela_growth": 95,
-    "most_recent_mcas_math_performance": "Not Meeting Expectations",
-    "most_recent_mcas_ela_performance": "Meeting Expectations",
-    "most_recent_mcas_math_scaled": 447,
-    "most_recent_mcas_ela_scaled": 490,
-    "most_recent_star_reading_percentile": 94,
-    "most_recent_star_math_percentile": 29,
-    "enrollment_status": "Active",
-    "date_of_birth": "2009-02-22T00:00:00.000Z",
-    "risk_level": 2,
-    "gender": "F",
-    "house": null,
-    "counselor": null,
-    "discipline_incidents_count": 0,
-    "absences_count": 0,
-    "tardies_count": 0,
-    "homeroom_name": "HEA 500",
-    "event_notes_without_restricted": [
-      {
-        "id": 9,
-        "student_id": 6,
-        "educator_id": 1,
-        "event_note_type_id": 302,
-        "text": "Engagement issues in ELA, sometimes shutting down altogether and refusing to engage in academic work.  Other times this is more disruptive, lashing out and yelling at other students and interfering with their learning.",
-        "recorded_at": "2010-10-17T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.080Z",
-        "updated_at": "2018-01-10T13:42:07.080Z",
-        "is_restricted": false
-      },
-      {
-        "id": 10,
-        "student_id": 6,
-        "educator_id": 1,
-        "event_note_type_id": 300,
-        "text": "Meetings with staff about lack of academic progress.",
-        "recorded_at": "2010-12-02T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.085Z",
-        "updated_at": "2018-01-10T13:42:07.085Z",
-        "is_restricted": false
-      },
-      {
-        "id": 11,
-        "student_id": 6,
-        "educator_id": 1,
-        "event_note_type_id": 304,
-        "text": "Engagement issues in ELA, sometimes shutting down altogether and refusing to engage in academic work.  Other times this is more disruptive, lashing out and yelling at other students and interfering with their learning.",
-        "recorded_at": "2011-01-01T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.089Z",
-        "updated_at": "2018-01-10T13:42:07.089Z",
-        "is_restricted": false
-      },
-      {
-        "id": 12,
-        "student_id": 6,
-        "educator_id": 1,
-        "event_note_type_id": 301,
-        "text": "Meredith is setting up a family meeting to discuss absences and tardies, and to follow up on concerns about supporting morning routines at home.",
-        "recorded_at": "2011-02-23T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.094Z",
-        "updated_at": "2018-01-10T13:42:07.094Z",
-        "is_restricted": false
-      },
-      {
-        "id": 13,
-        "student_id": 6,
-        "educator_id": 1,
-        "event_note_type_id": 302,
-        "text": "Not engaged in academic work and misbehaving in class.  There have been meetings between the teacher and support staff, and Sam has talked with Snow twice this week.  Philis following up with outside providers as well.",
-        "recorded_at": "2011-04-02T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.099Z",
-        "updated_at": "2018-01-10T13:42:07.099Z",
-        "is_restricted": false
-      },
-      {
-        "id": 14,
-        "student_id": 6,
-        "educator_id": 1,
-        "event_note_type_id": 301,
-        "text": "Meetings with staff about lack of academic progress.",
-        "recorded_at": "2011-06-01T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.104Z",
-        "updated_at": "2018-01-10T13:42:07.104Z",
-        "is_restricted": false
-      },
-      {
-        "id": 15,
-        "student_id": 6,
-        "educator_id": 1,
-        "event_note_type_id": 301,
-        "text": "Philis is meeting with parent next week and will present an attendance contract.",
-        "recorded_at": "2011-07-27T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.109Z",
-        "updated_at": "2018-01-10T13:42:07.109Z",
-        "is_restricted": false
-      },
-      {
-        "id": 16,
-        "student_id": 6,
-        "educator_id": 1,
-        "event_note_type_id": 300,
-        "text": "Snow will go through the special ed process for placement",
-        "recorded_at": "2011-08-29T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.113Z",
-        "updated_at": "2018-01-10T13:42:07.113Z",
-        "is_restricted": false
-      },
-      {
-        "id": 17,
-        "student_id": 6,
-        "educator_id": 1,
-        "event_note_type_id": 305,
-        "text": "Meetings with staff about lack of academic progress.",
-        "recorded_at": "2011-10-04T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:07.117Z",
-        "updated_at": "2018-01-10T13:42:07.117Z",
-        "is_restricted": false
-      }
-    ],
-    "interventions": [],
-    "sped_data": {
-      "sped_level": "—",
-      "sped_tooltip_message": null,
-      "sped_bubble_class": "sped"
-    },
-    "student_risk_level": {
-      "id": 6,
-      "student_id": 6,
-      "level": 2,
-      "created_at": "2018-01-10T13:42:56.261Z",
-      "updated_at": "2018-01-10T13:42:56.288Z",
-      "mcas_math_risk_level": null,
-      "star_math_risk_level": 2,
-      "mcas_ela_risk_level": null,
-      "star_reading_risk_level": 0,
-      "limited_english_proficiency_risk_level": null,
-      "explanation": {
-        "intro": "Snow is at Risk 2 because:",
-        "reasons": [
-          "Snow's STAR Math performance is in the 10-30 range."
-        ]
-      }
-    }
-  },
-  {
-    "id": 10,
-    "grade": "5",
-    "hispanic_latino": true,
-    "race": "White",
-    "free_reduced_lunch": "Free Lunch",
-    "created_at": "2018-01-10T13:42:08.713Z",
-    "updated_at": "2018-01-10T13:43:00.782Z",
-    "homeroom_id": 2,
-    "first_name": "Rapunzel",
-    "last_name": "Duck",
-    "state_id": null,
-    "home_language": "English",
-    "school_id": 2,
-    "registration_date": null,
-    "local_id": "000921",
-    "program_assigned": "2Way English",
-    "sped_placement": null,
-    "disability": null,
-    "sped_level_of_need": null,
-    "plan_504": "Not 504",
-    "limited_english_proficiency": "FLEP",
-    "most_recent_mcas_math_growth": 10,
-    "most_recent_mcas_ela_growth": 91,
-    "most_recent_mcas_math_performance": "Partially Meeting",
-    "most_recent_mcas_ela_performance": "Not Meeting Expectations",
-    "most_recent_mcas_math_scaled": 418,
-    "most_recent_mcas_ela_scaled": 502,
-    "most_recent_star_reading_percentile": 63,
-    "most_recent_star_math_percentile": 27,
-    "enrollment_status": "Active",
-    "date_of_birth": "2008-09-27T00:00:00.000Z",
-    "risk_level": 2,
-    "gender": "F",
-    "house": null,
-    "counselor": null,
-    "discipline_incidents_count": 0,
-    "absences_count": 3,
-    "tardies_count": 0,
-    "homeroom_name": "HEA 500",
-    "event_notes_without_restricted": [
-      {
-        "id": 38,
-        "student_id": 10,
-        "educator_id": 1,
-        "event_note_type_id": 301,
-        "text": "Meredith is setting up a family meeting to discuss absences and tardies, and to follow up on concerns about supporting morning routines at home.",
-        "recorded_at": "2010-10-07T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.778Z",
-        "updated_at": "2018-01-10T13:42:08.778Z",
-        "is_restricted": false
-      },
-      {
-        "id": 39,
-        "student_id": 10,
-        "educator_id": 1,
-        "event_note_type_id": 306,
-        "text": "Engagement issues in ELA, sometimes shutting down altogether and refusing to engage in academic work.  Other times this is more disruptive, lashing out and yelling at other students and interfering with their learning.",
-        "recorded_at": "2010-11-21T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.782Z",
-        "updated_at": "2018-01-10T13:42:08.782Z",
-        "is_restricted": false
-      },
-      {
-        "id": 40,
-        "student_id": 10,
-        "educator_id": 1,
-        "event_note_type_id": 300,
-        "text": "Meredith is setting up a family meeting to discuss absences and tardies, and to follow up on concerns about supporting morning routines at home.",
-        "recorded_at": "2011-01-02T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.787Z",
-        "updated_at": "2018-01-10T13:42:08.787Z",
-        "is_restricted": false
-      },
-      {
-        "id": 41,
-        "student_id": 10,
-        "educator_id": 1,
-        "event_note_type_id": 305,
-        "text": "Philis is meeting with parent next week and will present an attendance contract.",
-        "recorded_at": "2011-02-01T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.791Z",
-        "updated_at": "2018-01-10T13:42:08.791Z",
-        "is_restricted": false
-      },
-      {
-        "id": 42,
-        "student_id": 10,
-        "educator_id": 1,
-        "event_note_type_id": 305,
-        "text": "Meredith is setting up a family meeting to discuss absences and tardies, and to follow up on concerns about supporting morning routines at home.",
-        "recorded_at": "2011-03-06T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.795Z",
-        "updated_at": "2018-01-10T13:42:08.795Z",
-        "is_restricted": false
-      },
-      {
-        "id": 43,
-        "student_id": 10,
-        "educator_id": 1,
-        "event_note_type_id": 302,
-        "text": "Not engaged in academic work and misbehaving in class.  There have been meetings between the teacher and support staff, and Sam has talked with Rapunzel twice this week.  Philis following up with outside providers as well.",
-        "recorded_at": "2011-04-16T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.800Z",
-        "updated_at": "2018-01-10T13:42:08.800Z",
-        "is_restricted": false
-      },
-      {
-        "id": 44,
-        "student_id": 10,
-        "educator_id": 1,
-        "event_note_type_id": 306,
-        "text": "Not engaged in academic work and misbehaving in class.  There have been meetings between the teacher and support staff, and Sam has talked with Rapunzel twice this week.  Philis following up with outside providers as well.",
-        "recorded_at": "2011-06-14T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.804Z",
-        "updated_at": "2018-01-10T13:42:08.804Z",
-        "is_restricted": false
-      },
-      {
-        "id": 45,
-        "student_id": 10,
-        "educator_id": 1,
-        "event_note_type_id": 301,
-        "text": "Rapunzel will go through the special ed process for placement",
-        "recorded_at": "2011-08-06T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.809Z",
-        "updated_at": "2018-01-10T13:42:08.809Z",
-        "is_restricted": false
-      },
-      {
-        "id": 46,
-        "student_id": 10,
-        "educator_id": 1,
-        "event_note_type_id": 304,
-        "text": "Rapunzel will go through the special ed process for placement",
-        "recorded_at": "2011-09-26T00:00:00.000Z",
-        "created_at": "2018-01-10T13:42:08.813Z",
-        "updated_at": "2018-01-10T13:42:08.813Z",
-        "is_restricted": false
-      }
-    ],
-    "interventions": [],
-    "sped_data": {
-      "sped_level": "—",
-      "sped_tooltip_message": null,
-      "sped_bubble_class": "sped"
-    },
-    "student_risk_level": {
-      "id": 10,
-      "student_id": 10,
-      "level": 2,
-      "created_at": "2018-01-10T13:42:56.433Z",
-      "updated_at": "2018-01-10T13:42:56.459Z",
-      "mcas_math_risk_level": null,
-      "star_math_risk_level": 2,
-      "mcas_ela_risk_level": null,
-      "star_reading_risk_level": 1,
-      "limited_english_proficiency_risk_level": null,
-      "explanation": {
-        "intro": "Rapunzel is at Risk 2 because:",
-        "reasons": [
-          "Rapunzel's STAR Math performance is in the 10-30 range."
-        ]
-      }
-    }
-  }
-] // eslint-disable-line
