--- spec/javascripts/fixtures/schools.js
@@ -1,13 +1,0 @@
-export function healey() {
-  return {
-    local_id: 'HEA',
-    school_type: 'ESMS'
-  };
-}
-
-export function shs() {
-  return {
-    local_id: 'SHS',
-    school_type: 'HS'
-  };
-}
