--- spec/javascripts/fixtures/schools_overview_students.js
@@ -1,1 +1,0 @@
-export default [{"id":4,"grade":"5","hispanic_latino":true,"race":"White","free_reduced_lunch":"Free Lunch","created_at":"2018-01-10T13:42:06.268Z","updated_at":"2018-01-10T13:43:00.637Z","homeroom_id":2,"first_name":"Snow","last_name":"Skywalker","state_id":null,"home_language":"English","school_id":2,"registration_date":null,"local_id":"000209","program_assigned":"2Way Spanish","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Fluent","most_recent_mcas_math_growth":74,"most_recent_mcas_ela_growth":40,"most_recent_mcas_math_performance":"Exceeding Expectations","most_recent_mcas_ela_performance":"Meeting Expectations","most_recent_mcas_math_scaled":497,"most_recent_mcas_ela_scaled":499,"most_recent_star_reading_percentile":74,"most_recent_star_math_percentile":95,"enrollment_status":"Active","date_of_birth":"2009-05-03T00:00:00.000Z","risk_level":0,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":2,"tardies_count":0,"homeroom_name":"HEA 500","event_notes":[{"id":3,"student_id":4,"educator_id":1,"event_note_type_id":300,"recorded_at":"2010-10-31T00:00:00.000Z","is_restricted":false},{"id":4,"student_id":4,"educator_id":1,"event_note_type_id":306,"recorded_at":"2010-12-05T00:00:00.000Z","is_restricted":false},{"id":5,"student_id":4,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-01-08T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":7,"grade":"5","hispanic_latino":false,"race":"White","free_reduced_lunch":"Not Eligible","created_at":"2018-01-10T13:42:07.457Z","updated_at":"2018-01-10T13:43:00.710Z","homeroom_id":2,"first_name":"Chip","last_name":"Pan","state_id":null,"home_language":"English","school_id":2,"registration_date":null,"local_id":"000991","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"504","limited_english_proficiency":"FLEP","most_recent_mcas_math_growth":76,"most_recent_mcas_ela_growth":42,"most_recent_mcas_math_performance":"Partially Meeting","most_recent_mcas_ela_performance":"Meeting Expectations","most_recent_mcas_math_scaled":526,"most_recent_mcas_ela_scaled":512,"most_recent_star_reading_percentile":90,"most_recent_star_math_percentile":44,"enrollment_status":"Active","date_of_birth":"2009-03-12T00:00:00.000Z","risk_level":0,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":1,"tardies_count":0,"homeroom_name":"HEA 500","event_notes":[{"id":18,"student_id":7,"educator_id":1,"event_note_type_id":306,"recorded_at":"2010-10-30T00:00:00.000Z","is_restricted":false},{"id":19,"student_id":7,"educator_id":1,"event_note_type_id":305,"recorded_at":"2010-12-21T00:00:00.000Z","is_restricted":false},{"id":20,"student_id":7,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-02-16T00:00:00.000Z","is_restricted":false},{"id":21,"student_id":7,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-04-16T00:00:00.000Z","is_restricted":false},{"id":22,"student_id":7,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-05-26T00:00:00.000Z","is_restricted":false},{"id":23,"student_id":7,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-06-25T00:00:00.000Z","is_restricted":false},{"id":24,"student_id":7,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-08-19T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":9,"grade":"5","hispanic_latino":false,"race":"Black","free_reduced_lunch":"Free Lunch","created_at":"2018-01-10T13:42:08.285Z","updated_at":"2018-01-10T13:43:00.758Z","homeroom_id":2,"first_name":"Pluto","last_name":"Poppins","state_id":null,"home_language":"Haitian-Creole","school_id":2,"registration_date":null,"local_id":"000612","program_assigned":"2Way English","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Fluent","most_recent_mcas_math_growth":95,"most_recent_mcas_ela_growth":32,"most_recent_mcas_math_performance":"Not Meeting Expectations","most_recent_mcas_ela_performance":"Meeting Expectations","most_recent_mcas_math_scaled":448,"most_recent_mcas_ela_scaled":451,"most_recent_star_reading_percentile":54,"most_recent_star_math_percentile":100,"enrollment_status":"Active","date_of_birth":"2009-07-02T00:00:00.000Z","risk_level":0,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":2,"tardies_count":1,"homeroom_name":"HEA 500","event_notes":[{"id":30,"student_id":9,"educator_id":1,"event_note_type_id":300,"recorded_at":"2010-10-02T00:00:00.000Z","is_restricted":false},{"id":31,"student_id":9,"educator_id":1,"event_note_type_id":304,"recorded_at":"2010-11-26T00:00:00.000Z","is_restricted":false},{"id":32,"student_id":9,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-01-19T00:00:00.000Z","is_restricted":false},{"id":33,"student_id":9,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-03-08T00:00:00.000Z","is_restricted":false},{"id":34,"student_id":9,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-04-15T00:00:00.000Z","is_restricted":false},{"id":35,"student_id":9,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-06-05T00:00:00.000Z","is_restricted":false},{"id":36,"student_id":9,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-07-14T00:00:00.000Z","is_restricted":false},{"id":37,"student_id":9,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-09-02T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":12,"grade":"KF","hispanic_latino":false,"race":"Black","free_reduced_lunch":"Not Eligible","created_at":"2018-01-10T13:42:09.592Z","updated_at":"2018-01-10T13:43:00.843Z","homeroom_id":1,"first_name":"Winnie","last_name":"Kenobi","state_id":null,"home_language":"Portuguese","school_id":2,"registration_date":null,"local_id":"000406","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Fluent","most_recent_mcas_math_growth":80,"most_recent_mcas_ela_growth":14,"most_recent_mcas_math_performance":"Not Meeting Expectations","most_recent_mcas_ela_performance":"Meeting Expectations","most_recent_mcas_math_scaled":423,"most_recent_mcas_ela_scaled":554,"most_recent_star_reading_percentile":29,"most_recent_star_math_percentile":56,"enrollment_status":"Active","date_of_birth":"2013-11-08T00:00:00.000Z","risk_level":2,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":"HEA 003","event_notes":[{"id":54,"student_id":12,"educator_id":1,"event_note_type_id":305,"recorded_at":"2010-10-25T00:00:00.000Z","is_restricted":false},{"id":55,"student_id":12,"educator_id":1,"event_note_type_id":304,"recorded_at":"2010-12-06T00:00:00.000Z","is_restricted":false},{"id":56,"student_id":12,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-01-11T00:00:00.000Z","is_restricted":false},{"id":57,"student_id":12,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-02-21T00:00:00.000Z","is_restricted":false},{"id":58,"student_id":12,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-04-07T00:00:00.000Z","is_restricted":false},{"id":59,"student_id":12,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-05-20T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":14,"grade":"KF","hispanic_latino":false,"race":"Black","free_reduced_lunch":"Not Eligible","created_at":"2018-01-10T13:42:10.343Z","updated_at":"2018-01-10T13:43:00.900Z","homeroom_id":1,"first_name":"Rapunzel","last_name":"Poppins","state_id":null,"home_language":"English","school_id":2,"registration_date":null,"local_id":"000720","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP-Transitioning","most_recent_mcas_math_growth":80,"most_recent_mcas_ela_growth":52,"most_recent_mcas_math_performance":"Partially Meeting","most_recent_mcas_ela_performance":"Meeting Expectations","most_recent_mcas_math_scaled":561,"most_recent_mcas_ela_scaled":416,"most_recent_star_reading_percentile":94,"most_recent_star_math_percentile":45,"enrollment_status":"Active","date_of_birth":"2014-04-21T00:00:00.000Z","risk_level":0,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":"HEA 003","event_notes":[],"active_services":[],"summer_services":[],"interventions":[{"id":1,"name":"Behavior Contract","intervention_type_id":23,"comment":"The student had a tough day.","goal":"The student will pass the assessment.","start_date":"October 31, 2010","start_date_timestamp":"2010-10-31","end_date":"November 14, 2010","educator_email":"uri@demo.studentinsights.org","educator_id":1},{"id":2,"name":"ER Referral (Mental Health)","intervention_type_id":31,"comment":"Something happened.","goal":"The student will pass the assessment.","start_date":"December 10, 2010","start_date_timestamp":"2010-12-10","end_date":"December 24, 2010","educator_email":"uri@demo.studentinsights.org","educator_id":1}]},{"id":16,"grade":"KF","hispanic_latino":true,"race":"Black","free_reduced_lunch":"Not Eligible","created_at":"2018-01-10T13:42:11.087Z","updated_at":"2018-01-10T13:43:00.965Z","homeroom_id":1,"first_name":"Elsa","last_name":"Disney","state_id":null,"home_language":"Portuguese","school_id":2,"registration_date":null,"local_id":"000629","program_assigned":"Sp Ed","sped_placement":"Full Inclusion","disability":"Emotional","sped_level_of_need":"High","plan_504":"Not 504","limited_english_proficiency":"Fluent","most_recent_mcas_math_growth":25,"most_recent_mcas_ela_growth":27,"most_recent_mcas_math_performance":"Partially Meeting","most_recent_mcas_ela_performance":"P","most_recent_mcas_math_scaled":520,"most_recent_mcas_ela_scaled":249,"most_recent_star_reading_percentile":80,"most_recent_star_math_percentile":93,"enrollment_status":"Active","date_of_birth":"2013-10-21T00:00:00.000Z","risk_level":0,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":1,"tardies_count":0,"homeroom_name":"HEA 003","event_notes":[{"id":69,"student_id":16,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-10-10T00:00:00.000Z","is_restricted":false},{"id":70,"student_id":16,"educator_id":1,"event_note_type_id":301,"recorded_at":"2010-11-19T00:00:00.000Z","is_restricted":false},{"id":71,"student_id":16,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-12-26T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":17,"grade":"KF","hispanic_latino":true,"race":"Black","free_reduced_lunch":"Free Lunch","created_at":"2018-01-10T13:42:11.443Z","updated_at":"2018-01-10T13:43:00.991Z","homeroom_id":1,"first_name":"Pluto","last_name":"Pan","state_id":null,"home_language":"Haitian-Creole","school_id":2,"registration_date":null,"local_id":"000698","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP","most_recent_mcas_math_growth":31,"most_recent_mcas_ela_growth":54,"most_recent_mcas_math_performance":"Meeting Expectations","most_recent_mcas_ela_performance":"Partially Meeting","most_recent_mcas_math_scaled":422,"most_recent_mcas_ela_scaled":598,"most_recent_star_reading_percentile":9,"most_recent_star_math_percentile":38,"enrollment_status":"Active","date_of_birth":"2014-06-04T00:00:00.000Z","risk_level":3,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":"HEA 003","event_notes":[{"id":72,"student_id":17,"educator_id":1,"event_note_type_id":306,"recorded_at":"2010-10-02T00:00:00.000Z","is_restricted":false},{"id":73,"student_id":17,"educator_id":1,"event_note_type_id":306,"recorded_at":"2010-11-27T00:00:00.000Z","is_restricted":false},{"id":74,"student_id":17,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-01-18T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":19,"grade":"KF","hispanic_latino":true,"race":"White","free_reduced_lunch":"Not Eligible","created_at":"2018-01-10T13:42:12.200Z","updated_at":"2018-01-10T13:43:01.045Z","homeroom_id":1,"first_name":"Winnie","last_name":"White","state_id":null,"home_language":"English","school_id":2,"registration_date":null,"local_id":"000573","program_assigned":"2Way Spanish","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Fluent","most_recent_mcas_math_growth":41,"most_recent_mcas_ela_growth":4,"most_recent_mcas_math_performance":"Meeting Expectations","most_recent_mcas_ela_performance":"Not Meeting Expectations","most_recent_mcas_math_scaled":497,"most_recent_mcas_ela_scaled":540,"most_recent_star_reading_percentile":78,"most_recent_star_math_percentile":96,"enrollment_status":"Active","date_of_birth":"2014-01-06T00:00:00.000Z","risk_level":0,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":"HEA 003","event_notes":[{"id":84,"student_id":19,"educator_id":1,"event_note_type_id":300,"recorded_at":"2010-10-27T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":3,"grade":"5","hispanic_latino":true,"race":"White","free_reduced_lunch":"Free Lunch","created_at":"2018-01-10T13:42:05.749Z","updated_at":"2018-01-10T13:43:00.612Z","homeroom_id":2,"first_name":"Aladdin","last_name":"Kenobi","state_id":null,"home_language":"Haitian-Creole","school_id":2,"registration_date":null,"local_id":"00085","program_assigned":"2Way Spanish","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP","most_recent_mcas_math_growth":33,"most_recent_mcas_ela_growth":2,"most_recent_mcas_math_performance":"P","most_recent_mcas_ela_performance":"Meeting Expectations","most_recent_mcas_math_scaled":255,"most_recent_mcas_ela_scaled":549,"most_recent_star_reading_percentile":76,"most_recent_star_math_percentile":2,"enrollment_status":"Active","date_of_birth":"2009-08-13T00:00:00.000Z","risk_level":3,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":"HEA 500","event_notes":[{"id":1,"student_id":3,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-10-15T00:00:00.000Z","is_restricted":false},{"id":2,"student_id":3,"educator_id":1,"event_note_type_id":304,"recorded_at":"2010-12-10T00:00:00.000Z","is_restricted":false}],"active_services":[{"id":62,"student_id":3,"provided_by_educator_name":"Teacher, Sarah","service_type_id":503,"date_started":"2018-01-20T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":1,"recorded_at":"2018-01-20T17:06:27.404Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":63,"student_id":3,"provided_by_educator_name":"Teacher, Sarah","service_type_id":503,"date_started":"2018-01-20T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":1,"recorded_at":"2018-01-20T17:06:35.852Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"summer_services":[],"interventions":[]},{"id":58,"grade":"3","hispanic_latino":false,"race":"Asian","free_reduced_lunch":"Free Lunch","created_at":"2018-01-10T13:42:27.990Z","updated_at":"2018-01-10T13:43:02.043Z","homeroom_id":null,"first_name":"Winnie","last_name":"Poppins","state_id":null,"home_language":"Spanish","school_id":2,"registration_date":null,"local_id":"000282","program_assigned":"Sp Ed","sped_placement":"Private Separate","disability":"Autism","sped_level_of_need":"Low >= 2","plan_504":"Not 504","limited_english_proficiency":"FLEP-Transitioning","most_recent_mcas_math_growth":19,"most_recent_mcas_ela_growth":56,"most_recent_mcas_math_performance":"P","most_recent_mcas_ela_performance":"A","most_recent_mcas_math_scaled":280,"most_recent_mcas_ela_scaled":237,"most_recent_star_reading_percentile":63,"most_recent_star_math_percentile":74,"enrollment_status":"Active","date_of_birth":"2011-07-29T00:00:00.000Z","risk_level":0,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":1,"tardies_count":1,"homeroom_name":null,"event_notes":[{"id":246,"student_id":58,"educator_id":1,"event_note_type_id":301,"recorded_at":"2010-10-02T00:00:00.000Z","is_restricted":false},{"id":247,"student_id":58,"educator_id":1,"event_note_type_id":304,"recorded_at":"2010-11-28T00:00:00.000Z","is_restricted":false},{"id":248,"student_id":58,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-01-02T00:00:00.000Z","is_restricted":false},{"id":249,"student_id":58,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-02-04T00:00:00.000Z","is_restricted":false},{"id":250,"student_id":58,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-03-28T00:00:00.000Z","is_restricted":false},{"id":251,"student_id":58,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-05-24T00:00:00.000Z","is_restricted":false},{"id":252,"student_id":58,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-07-18T00:00:00.000Z","is_restricted":false},{"id":253,"student_id":58,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-08-17T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":59,"grade":"3","hispanic_latino":true,"race":"Asian","free_reduced_lunch":"Not Eligible","created_at":"2018-01-10T13:42:28.546Z","updated_at":"2018-01-10T13:43:02.079Z","homeroom_id":null,"first_name":"Mowgli","last_name":"White","state_id":null,"home_language":"English","school_id":2,"registration_date":null,"local_id":"000416","program_assigned":"2Way English","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Fluent","most_recent_mcas_math_growth":37,"most_recent_mcas_ela_growth":15,"most_recent_mcas_math_performance":"Not Meeting Expectations","most_recent_mcas_ela_performance":"P","most_recent_mcas_math_scaled":464,"most_recent_mcas_ela_scaled":267,"most_recent_star_reading_percentile":49,"most_recent_star_math_percentile":100,"enrollment_status":"Active","date_of_birth":"2011-06-16T00:00:00.000Z","risk_level":0,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":4,"tardies_count":0,"homeroom_name":null,"event_notes":[{"id":254,"student_id":59,"educator_id":1,"event_note_type_id":300,"recorded_at":"2010-10-17T00:00:00.000Z","is_restricted":false},{"id":255,"student_id":59,"educator_id":1,"event_note_type_id":300,"recorded_at":"2010-11-27T00:00:00.000Z","is_restricted":false},{"id":256,"student_id":59,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-01-24T00:00:00.000Z","is_restricted":false},{"id":257,"student_id":59,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-03-24T00:00:00.000Z","is_restricted":false},{"id":258,"student_id":59,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-05-11T00:00:00.000Z","is_restricted":false},{"id":259,"student_id":59,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-07-04T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":60,"grade":"3","hispanic_latino":false,"race":"Black","free_reduced_lunch":"Not Eligible","created_at":"2018-01-10T13:42:29.016Z","updated_at":"2018-01-10T13:43:02.105Z","homeroom_id":null,"first_name":"Snow","last_name":"Skywalker","state_id":null,"home_language":"Spanish","school_id":2,"registration_date":null,"local_id":"000900","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Fluent","most_recent_mcas_math_growth":40,"most_recent_mcas_ela_growth":32,"most_recent_mcas_math_performance":"W","most_recent_mcas_ela_performance":"P","most_recent_mcas_math_scaled":201,"most_recent_mcas_ela_scaled":221,"most_recent_star_reading_percentile":76,"most_recent_star_math_percentile":7,"enrollment_status":"Active","date_of_birth":"2011-07-21T00:00:00.000Z","risk_level":3,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":1,"homeroom_name":null,"event_notes":[{"id":260,"student_id":60,"educator_id":1,"event_note_type_id":306,"recorded_at":"2010-10-06T00:00:00.000Z","is_restricted":false},{"id":261,"student_id":60,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-11-27T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":63,"grade":"3","hispanic_latino":false,"race":"Black","free_reduced_lunch":"Not Eligible","created_at":"2018-01-10T13:42:30.545Z","updated_at":"2018-01-10T13:43:02.181Z","homeroom_id":null,"first_name":"Mickey","last_name":"Kenobi","state_id":null,"home_language":"Spanish","school_id":2,"registration_date":null,"local_id":"000897","program_assigned":"2Way English","sped_placement":null,"disability":"Communication","sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP","most_recent_mcas_math_growth":7,"most_recent_mcas_ela_growth":98,"most_recent_mcas_math_performance":"W","most_recent_mcas_ela_performance":"Exceeding Expectations","most_recent_mcas_math_scaled":258,"most_recent_mcas_ela_scaled":526,"most_recent_star_reading_percentile":33,"most_recent_star_math_percentile":49,"enrollment_status":"Active","date_of_birth":"2010-10-10T00:00:00.000Z","risk_level":3,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":null,"event_notes":[{"id":266,"student_id":63,"educator_id":1,"event_note_type_id":304,"recorded_at":"2010-10-28T00:00:00.000Z","is_restricted":false}],"active_services":[{"id":64,"student_id":63,"provided_by_educator_name":"Teacher, Sarah","service_type_id":504,"date_started":"2018-01-24T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":1,"recorded_at":"2018-01-24T02:17:23.267Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":65,"student_id":63,"provided_by_educator_name":"Teacher, Sarah","service_type_id":504,"date_started":"2018-01-24T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":1,"recorded_at":"2018-01-24T02:17:38.394Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"summer_services":[],"interventions":[]},{"id":64,"grade":"3","hispanic_latino":true,"race":"White","free_reduced_lunch":"Not Eligible","created_at":"2018-01-10T13:42:30.929Z","updated_at":"2018-01-10T13:43:02.203Z","homeroom_id":null,"first_name":"Pluto","last_name":"Skywalker","state_id":null,"home_language":"Portuguese","school_id":2,"registration_date":null,"local_id":"00093","program_assigned":"Sp Ed","sped_placement":"Private Separate","disability":"Autism","sped_level_of_need":"Low < 2","plan_504":"Not 504","limited_english_proficiency":"FLEP-Transitioning","most_recent_mcas_math_growth":26,"most_recent_mcas_ela_growth":18,"most_recent_mcas_math_performance":"NI","most_recent_mcas_ela_performance":"Not Meeting Expectations","most_recent_mcas_math_scaled":234,"most_recent_mcas_ela_scaled":473,"most_recent_star_reading_percentile":21,"most_recent_star_math_percentile":24,"enrollment_status":"Active","date_of_birth":"2010-10-02T00:00:00.000Z","risk_level":2,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":null,"event_notes":[{"id":267,"student_id":64,"educator_id":1,"event_note_type_id":306,"recorded_at":"2010-10-04T00:00:00.000Z","is_restricted":false},{"id":268,"student_id":64,"educator_id":1,"event_note_type_id":300,"recorded_at":"2010-11-29T00:00:00.000Z","is_restricted":false},{"id":269,"student_id":64,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-01-15T00:00:00.000Z","is_restricted":false},{"id":270,"student_id":64,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-03-12T00:00:00.000Z","is_restricted":false},{"id":271,"student_id":64,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-05-02T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":65,"grade":"3","hispanic_latino":false,"race":"White","free_reduced_lunch":"Free Lunch","created_at":"2018-01-10T13:42:31.311Z","updated_at":"2018-01-10T13:43:02.226Z","homeroom_id":null,"first_name":"Daisy","last_name":"Mouse","state_id":null,"home_language":"English","school_id":2,"registration_date":null,"local_id":"000944","program_assigned":"Reg Ed","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP-Transitioning","most_recent_mcas_math_growth":28,"most_recent_mcas_ela_growth":13,"most_recent_mcas_math_performance":"P","most_recent_mcas_ela_performance":"P","most_recent_mcas_math_scaled":203,"most_recent_mcas_ela_scaled":251,"most_recent_star_reading_percentile":16,"most_recent_star_math_percentile":92,"enrollment_status":"Active","date_of_birth":"2010-12-16T00:00:00.000Z","risk_level":2,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":2,"tardies_count":1,"homeroom_name":null,"event_notes":[{"id":272,"student_id":65,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-10-26T00:00:00.000Z","is_restricted":false},{"id":273,"student_id":65,"educator_id":1,"event_note_type_id":301,"recorded_at":"2010-12-10T00:00:00.000Z","is_restricted":false},{"id":274,"student_id":65,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-01-17T00:00:00.000Z","is_restricted":false},{"id":275,"student_id":65,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-03-03T00:00:00.000Z","is_restricted":false},{"id":276,"student_id":65,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-04-08T00:00:00.000Z","is_restricted":false},{"id":277,"student_id":65,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-06-04T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":57,"grade":"3","hispanic_latino":true,"race":"Asian","free_reduced_lunch":"Free Lunch","created_at":"2018-01-10T13:42:27.407Z","updated_at":"2018-01-10T13:43:02.018Z","homeroom_id":null,"first_name":"Minnie","last_name":"White","state_id":null,"home_language":"Portuguese","school_id":2,"registration_date":null,"local_id":"000815","program_assigned":"Sp Ed","sped_placement":"Private Separate","disability":"Specific LDs","sped_level_of_need":"Low >= 2","plan_504":"Not 504","limited_english_proficiency":"FLEP","most_recent_mcas_math_growth":11,"most_recent_mcas_ela_growth":50,"most_recent_mcas_math_performance":"Partially Meeting","most_recent_mcas_ela_performance":"P","most_recent_mcas_math_scaled":401,"most_recent_mcas_ela_scaled":215,"most_recent_star_reading_percentile":43,"most_recent_star_math_percentile":45,"enrollment_status":"Active","date_of_birth":"2010-11-08T00:00:00.000Z","risk_level":1,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":null,"event_notes":[{"id":242,"student_id":57,"educator_id":1,"event_note_type_id":301,"recorded_at":"2010-10-28T00:00:00.000Z","is_restricted":false},{"id":243,"student_id":57,"educator_id":1,"event_note_type_id":306,"recorded_at":"2010-12-15T00:00:00.000Z","is_restricted":false},{"id":244,"student_id":57,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-02-08T00:00:00.000Z","is_restricted":false},{"id":245,"student_id":57,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-03-11T00:00:00.000Z","is_restricted":false}],"active_services":[{"id":28,"student_id":57,"provided_by_educator_name":null,"service_type_id":510,"date_started":"2017-11-28T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":12,"recorded_at":"2017-12-01T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":29,"student_id":57,"provided_by_educator_name":"Disney, Uri","service_type_id":504,"date_started":"2017-10-08T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":4,"recorded_at":"2017-10-08T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":30,"student_id":57,"provided_by_educator_name":"Disney, Uri","service_type_id":513,"date_started":"2017-08-11T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":4,"recorded_at":"2017-08-13T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":31,"student_id":57,"provided_by_educator_name":"Disney, Uri","service_type_id":505,"date_started":"2017-06-15T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":6,"recorded_at":"2017-06-21T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":32,"student_id":57,"provided_by_educator_name":null,"service_type_id":506,"date_started":"2017-04-21T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":8,"recorded_at":"2017-04-24T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"summer_services":[{"id":28,"student_id":57,"provided_by_educator_name":null,"service_type_id":510,"date_started":"2017-11-28T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":12,"recorded_at":"2017-12-01T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"interventions":[]},{"id":6,"grade":"5","hispanic_latino":false,"race":"Asian","free_reduced_lunch":"Free Lunch","created_at":"2018-01-10T13:42:07.064Z","updated_at":"2018-01-10T13:43:00.686Z","homeroom_id":2,"first_name":"Snow","last_name":"Kenobi","state_id":null,"home_language":"English","school_id":2,"registration_date":null,"local_id":"000407","program_assigned":"2Way English","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP-Transitioning","most_recent_mcas_math_growth":0,"most_recent_mcas_ela_growth":95,"most_recent_mcas_math_performance":"Not Meeting Expectations","most_recent_mcas_ela_performance":"Meeting Expectations","most_recent_mcas_math_scaled":447,"most_recent_mcas_ela_scaled":490,"most_recent_star_reading_percentile":94,"most_recent_star_math_percentile":29,"enrollment_status":"Active","date_of_birth":"2009-02-22T00:00:00.000Z","risk_level":2,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":"HEA 500","event_notes":[{"id":9,"student_id":6,"educator_id":1,"event_note_type_id":302,"recorded_at":"2010-10-17T00:00:00.000Z","is_restricted":false},{"id":10,"student_id":6,"educator_id":1,"event_note_type_id":300,"recorded_at":"2010-12-02T00:00:00.000Z","is_restricted":false},{"id":11,"student_id":6,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-01-01T00:00:00.000Z","is_restricted":false},{"id":12,"student_id":6,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-02-23T00:00:00.000Z","is_restricted":false},{"id":13,"student_id":6,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-04-02T00:00:00.000Z","is_restricted":false},{"id":14,"student_id":6,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-06-01T00:00:00.000Z","is_restricted":false},{"id":15,"student_id":6,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-07-27T00:00:00.000Z","is_restricted":false},{"id":16,"student_id":6,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-08-29T00:00:00.000Z","is_restricted":false},{"id":17,"student_id":6,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-10-04T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":10,"grade":"5","hispanic_latino":true,"race":"White","free_reduced_lunch":"Free Lunch","created_at":"2018-01-10T13:42:08.713Z","updated_at":"2018-01-10T13:43:00.782Z","homeroom_id":2,"first_name":"Rapunzel","last_name":"Duck","state_id":null,"home_language":"English","school_id":2,"registration_date":null,"local_id":"000921","program_assigned":"2Way English","sped_placement":null,"disability":null,"sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP","most_recent_mcas_math_growth":10,"most_recent_mcas_ela_growth":91,"most_recent_mcas_math_performance":"Partially Meeting","most_recent_mcas_ela_performance":"Not Meeting Expectations","most_recent_mcas_math_scaled":418,"most_recent_mcas_ela_scaled":502,"most_recent_star_reading_percentile":63,"most_recent_star_math_percentile":27,"enrollment_status":"Active","date_of_birth":"2008-09-27T00:00:00.000Z","risk_level":2,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":3,"tardies_count":0,"homeroom_name":"HEA 500","event_notes":[{"id":38,"student_id":10,"educator_id":1,"event_note_type_id":301,"recorded_at":"2010-10-07T00:00:00.000Z","is_restricted":false},{"id":39,"student_id":10,"educator_id":1,"event_note_type_id":306,"recorded_at":"2010-11-21T00:00:00.000Z","is_restricted":false},{"id":40,"student_id":10,"educator_id":1,"event_note_type_id":300,"recorded_at":"2011-01-02T00:00:00.000Z","is_restricted":false},{"id":41,"student_id":10,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-02-01T00:00:00.000Z","is_restricted":false},{"id":42,"student_id":10,"educator_id":1,"event_note_type_id":305,"recorded_at":"2011-03-06T00:00:00.000Z","is_restricted":false},{"id":43,"student_id":10,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-04-16T00:00:00.000Z","is_restricted":false},{"id":44,"student_id":10,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-06-14T00:00:00.000Z","is_restricted":false},{"id":45,"student_id":10,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-08-06T00:00:00.000Z","is_restricted":false},{"id":46,"student_id":10,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-09-26T00:00:00.000Z","is_restricted":false},{"id":514,"student_id":10,"educator_id":1,"event_note_type_id":300,"recorded_at":"2018-01-27T17:14:08.065Z","is_restricted":true},{"id":515,"student_id":10,"educator_id":1,"event_note_type_id":305,"recorded_at":"2018-01-27T17:14:13.318Z","is_restricted":true},{"id":516,"student_id":10,"educator_id":1,"event_note_type_id":306,"recorded_at":"2018-01-27T17:14:16.006Z","is_restricted":true}],"active_services":[],"summer_services":[],"interventions":[]},{"id":13,"grade":"KF","hispanic_latino":false,"race":"Black","free_reduced_lunch":"Not Eligible","created_at":"2018-01-10T13:42:09.966Z","updated_at":"2018-01-10T13:43:00.873Z","homeroom_id":1,"first_name":"Rapunzel","last_name":"Poppins","state_id":null,"home_language":"Haitian-Creole","school_id":2,"registration_date":null,"local_id":"000366","program_assigned":"2Way Spanish","sped_placement":null,"disability":"Communication","sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"FLEP","most_recent_mcas_math_growth":83,"most_recent_mcas_ela_growth":95,"most_recent_mcas_math_performance":"NI","most_recent_mcas_ela_performance":"Partially Meeting","most_recent_mcas_math_scaled":226,"most_recent_mcas_ela_scaled":581,"most_recent_star_reading_percentile":100,"most_recent_star_math_percentile":43,"enrollment_status":"Active","date_of_birth":"2014-05-09T00:00:00.000Z","risk_level":2,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":0,"homeroom_name":"HEA 003","event_notes":[{"id":60,"student_id":13,"educator_id":1,"event_note_type_id":300,"recorded_at":"2010-10-07T00:00:00.000Z","is_restricted":false},{"id":61,"student_id":13,"educator_id":1,"event_note_type_id":301,"recorded_at":"2010-11-23T00:00:00.000Z","is_restricted":false},{"id":62,"student_id":13,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-01-13T00:00:00.000Z","is_restricted":false},{"id":63,"student_id":13,"educator_id":1,"event_note_type_id":306,"recorded_at":"2011-02-19T00:00:00.000Z","is_restricted":false},{"id":64,"student_id":13,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-04-09T00:00:00.000Z","is_restricted":false},{"id":65,"student_id":13,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-06-07T00:00:00.000Z","is_restricted":false},{"id":66,"student_id":13,"educator_id":1,"event_note_type_id":302,"recorded_at":"2011-07-19T00:00:00.000Z","is_restricted":false},{"id":67,"student_id":13,"educator_id":1,"event_note_type_id":301,"recorded_at":"2011-09-01T00:00:00.000Z","is_restricted":false},{"id":68,"student_id":13,"educator_id":1,"event_note_type_id":304,"recorded_at":"2011-10-16T00:00:00.000Z","is_restricted":false}],"active_services":[],"summer_services":[],"interventions":[]},{"id":20,"grade":"KF","hispanic_latino":false,"race":"White","free_reduced_lunch":"Free Lunch","created_at":"2018-01-10T13:42:12.556Z","updated_at":"2018-01-10T13:43:01.071Z","homeroom_id":1,"first_name":"Snow","last_name":"White","state_id":null,"home_language":"Haitian-Creole","school_id":2,"registration_date":null,"local_id":"000687","program_assigned":"Reg Ed","sped_placement":null,"disability":"Specific LDs","sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Fluent","most_recent_mcas_math_growth":55,"most_recent_mcas_ela_growth":86,"most_recent_mcas_math_performance":null,"most_recent_mcas_ela_performance":"Meeting Expectations","most_recent_mcas_math_scaled":269,"most_recent_mcas_ela_scaled":556,"most_recent_star_reading_percentile":16,"most_recent_star_math_percentile":12,"enrollment_status":"Active","date_of_birth":"2014-03-03T00:00:00.000Z","risk_level":2,"gender":"F","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":0,"tardies_count":2,"homeroom_name":"HEA 003","event_notes":[{"id":85,"student_id":20,"educator_id":1,"event_note_type_id":304,"recorded_at":"2010-10-12T00:00:00.000Z","is_restricted":false}],"active_services":[{"id":1,"student_id":20,"provided_by_educator_name":"Disney, Uri","service_type_id":514,"date_started":"2017-11-12T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":9,"recorded_at":"2017-11-19T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":2,"student_id":20,"provided_by_educator_name":null,"service_type_id":506,"date_started":"2017-09-18T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":9,"recorded_at":"2017-09-24T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"summer_services":[],"interventions":[]},{"id":61,"grade":"3","hispanic_latino":false,"race":"Black","free_reduced_lunch":"Not Eligible","created_at":"2018-01-10T13:42:29.467Z","updated_at":"2018-01-10T13:43:02.132Z","homeroom_id":null,"first_name":"Elsa","last_name":"Poppins","state_id":null,"home_language":"Portuguese","school_id":2,"registration_date":null,"local_id":"000758","program_assigned":"2Way English","sped_placement":null,"disability":"Specific LDs","sped_level_of_need":null,"plan_504":"Not 504","limited_english_proficiency":"Fluent","most_recent_mcas_math_growth":51,"most_recent_mcas_ela_growth":87,"most_recent_mcas_math_performance":"Partially Meeting","most_recent_mcas_ela_performance":"NI","most_recent_mcas_math_scaled":498,"most_recent_mcas_ela_scaled":251,"most_recent_star_reading_percentile":72,"most_recent_star_math_percentile":44,"enrollment_status":"Active","date_of_birth":"2010-12-11T00:00:00.000Z","risk_level":2,"gender":"M","house":null,"counselor":null,"discipline_incidents_count":0,"absences_count":2,"tardies_count":0,"homeroom_name":null,"event_notes":[],"active_services":[{"id":33,"student_id":61,"provided_by_educator_name":"Disney, Uri","service_type_id":509,"date_started":"2017-11-26T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":8,"recorded_at":"2017-12-01T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":34,"student_id":61,"provided_by_educator_name":"Disney, Uri","service_type_id":502,"date_started":"2017-10-28T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":7,"recorded_at":"2017-10-29T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":35,"student_id":61,"provided_by_educator_name":null,"service_type_id":508,"date_started":"2017-09-07T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":10,"recorded_at":"2017-09-10T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":36,"student_id":61,"provided_by_educator_name":null,"service_type_id":513,"date_started":"2017-08-05T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":6,"recorded_at":"2017-08-10T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"summer_services":[{"id":33,"student_id":61,"provided_by_educator_name":"Disney, Uri","service_type_id":509,"date_started":"2017-11-26T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":8,"recorded_at":"2017-12-01T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"interventions":[]}];
